Convex Hull

DelaunayTriangulation.convex_hull — Function

convex_hull(points; IntegerType::Type{I}=Int) where {I} -> ConvexHull

Computes the convex hull of points. The monotone chain algorithm is used.

Arguments

points: The set of points.

Output

ch: The ConvexHull.

DelaunayTriangulation.convex_hull! — Function

convex_hull!(tri::Triangulation; reconstruct=has_boundary_nodes(tri))

Updates the convex_hull field of tri to match the current triangulation.

Arguments

tri::Triangulation: The Triangulation.

Keyword Arguments

reconstruct=has_boundary_nodes(tri): If true, then the convex hull is reconstructed from scratch, using convex_hull on the points. Otherwise, computes the convex hull using the ghost triangles of tri. If there are no ghost triangles but reconstruct=true, then the convex hull is reconstructed from scratch.

convex_hull!(ch::ConvexHull{P,I}) where {P,I}

Using the points in ch, computes the convex hull in-place.

See also convex_hull.