Representative Coordinates

When we are building the triangulations, we need to have a point that represents the center of the domain at each stage so that ghost triangles can be worked with for point location. The main type we use for this is RepresentativeCoordinates:

DelaunayTriangulation.RepresentativeCoordinates — Type

mutable struct RepresentativeCoordinates{I,T}

Struct representing the representative point of a region.

See also RepresentativePointList.

Fields

x: The x-coordinate.

y: The y-coordinate.

n: The number of points defining the centroid, if that is what is being used.

Usage

See reset!, add_point!, delete_point!, and compute_representative_point!.

More accurately, we maintain a list of such coordinates, one for each boundary curve. This list is a const, given below.

DelaunayTriangulation.RepresentativePointList — Constant

const RepresentativePointList = Dict{Int64,RepresentativeCoordinates{Int64, Float64}}()

This is a list of points for a set of representative points corresponding to multiple curves.

See also update_centroid_after_addition, update_centroid_after_deletion, reset_centroids!, empty_centroids!, get_representative_point_coordinates, and new_centroid.

Warning

While the function compute_representative_points! computes an appropriate visual center of the polygon represented by the curves, i.e. by joining points, the update functions like update_centroid_after_addition and update_centroid_after_deletion treat the centroid as if it is a mean of all the points. If you need to use the actual visual centers of the polygons, you will need to reuse compute_representative_points! again (as is done at the end of triangulate).

This list is updated at the end of a triangulation to be given by the poles of inaccessibilities (see the sidebar), but when the triangulation is being built it is given by the arithmetic mean of all points in the triangulation. Some other docstrings are below.

DelaunayTriangulation.get_representative_point_coordinates — Function

get_representative_point_coordinates(i::I, ::Type{T}) where {I, T}

Given an index i and some number type T, returns the coordinates corresponding to the centroid in RepresentativePointList at the key i. The returned value is a Tuple (T(x), T(y)) of the coordinates.

DelaunayTriangulation.compute_representative_points! — Method

compute_representative_points!(points, boundary_nodes; precision = 1.0)

Given a list of points and a list of boundary_nodes, computes visual centers for the polygons represented by these curves using pole_of_inaccessibility. The keyword argument precision is the precision used in pole_of_inaccessibility.

See also RepresentativePointList.

Warning

While this function computes an appropriate visual center of the polygon represented by the curves, i.e. by joining points, the update functions like update_centroid_after_addition and update_centroid_after_deletion treat the centroid as if it is a mean of all the points. If you need to use the actual visual centers of the polygons, you will need to reuse this function again (as is done at the end of triangulate).