Triangulation

The most important structure of the package is the Triangulation data structure. Its complete definition is given below, and then afterwards we give all the docstrings for methods that are useful for working with triangulations.

DelaunayTriangulation.Triangulation — Type

Triangulation{P,Ts,I,E,Es,BN,BNM,B,BIR,BPL}

Struct representing a Delaunay triangulation, as constructed via e.g. triangulate or generate_mesh.

Fields

points::P

The points used to construct the triangulation. Do note that this is not necessarly the same as the set of points appearing in the triangulation, as could happen e.g. if some points were deleted from the triangulation. If that is what you need, then see each_vertex.

triangles::Ts

The triangles in the triangulation, all given with positive orientation. This field can include ghost triangles. If you don't want ghost triangles, see each_solid_triangle.

adjacent::Adjacent{I, E}

The Adjacent map.

adjacent2vertex::Adjacent2Vertex{I,Es,E}

The Adjacent2Vertex map.

graph::Graph{I}

The Graph.

boundary_nodes::BN

The boundary nodes in the triangulation that you provide, matching the specification given in the documentation.

boundary_edge_map::BNM

This is a Dict that maps all of the boundary edges to their position in boundary_nodes. See also construct_boundary_edge_map.

boundary_index_ranges::BIR

This is an OrderedDict that maps a boundary index to a range of all other boundary indices that the corresponding boundary curve could correspond to. For example, for a curve with four segments, there are four possible boundary indices that a segment could correspond to. See also construct_boundary_index_ranges.

constrained_edges::Es

This is the set of extra edges added into the triangulation that you have provided. This will not include any of the other constrained edges from boundary_nodes.

all_constrained_edges::Es

This is the set of all constrained edges appearing in the triangulation, essentially given as the union of constrained_edges and boundary_nodes. This is different from the constrained_edges field as it contains edges from both of these fields, e.g. there might be an edge in constrained_edges that is not yet in the triangulation, but it will definitely not appear in all_constrained_edges.

convex_hull::ConvexHull{P,Vector{I}}

The ConvexHull of points. See also convex_hull.

representative_point_list::RPL

The Dict of points giving representative points for each boundary curve, or for the convex hull if boundary_nodes is empty. These representative points are used for interpreting ghost vertices.

Constructors

There are several ways to construct this struct directly, although in most cases you should be using triangulate or generate_mesh.

Default Constructor

The default constructor is available, i.e.

Triangulation(
    points,
    triangles,
    adjacent,
    adjacent2vertex,
    graph,
    boundary_nodes,
    boundary_edge_map,
    boundary_map,
    boundary_index_ranges,
    constrained_edges,
    all_constrained_edges,
    convex_hull,
    representative_point_list
)

Empty Triangulation

An empty triangulation can be initalised with the following method,

Triangulation(points;
    IntegerType=Int,
    EdgeType=NTuple{2,IntegerType},
    TriangleType=NTuple{3,IntegerType},
    EdgesType=Set{EdgeType},
    TrianglesType=Set{TriangleType},
    boundary_nodes=IntegerType[],
    constrained_edges=initialise_edges(EdgesType),
    representative_point_list=get_empty_representative_points(IntegerType, number_type(points))
)

Triangulation From an Existing Mesh

A method is available from constructing a mesh from an existing set of points, triangles, and boundary nodes, mainly existing for the purpose of generate_mesh:

Triangulation(points, triangles, boundary_nodes;
    IntegerType=Int,
    EdgeType=NTuple{2,IntegerType},
    TriangleType=NTuple{3,IntegerType},
    EdgesType=Set{EdgeType},
    TrianglesType=Set{TriangleType},
    add_ghost_triangles=false
)

with add_ghost_triangles calling add_ghost_triangles! at the end of the constructor.

Wrapping a Triangulation with constrainededgepoints

This method is used in triangulate for wrapping a triangulation with a set of constraints:

remake_triangulation_with_constraints(triangulation, edges, boundary_nodes)

which returns (bn_map, bn_range, tri), where bn_map is the boundary_map that isn't yet added, bn_range is the boundary_index_range that isn't yet added either, and tri is the wrapped triangulation that includes the constrained edges and the boundary_nodes. Note that either edges or boundary_nodes can be nothing.

This can be used together with the method

replace_boundary_dict_information(triangulation, bn_map, bn_range)

which returns a new triangulation with the boundary_map and boundary_index_range replaced, completing the wrapper.

Each field has its own accessor:

DelaunayTriangulation.get_points — Method

get_points(tri::Triangulation)

Returns the points field from the triangulation tri.

DelaunayTriangulation.get_triangles — Method

get_triangles(tri::Triangulation)

Returns the triangles field from the triangulation tri.

DelaunayTriangulation.get_adjacent — Method

get_adjacent(tri::Triangulation)

Returns the adjacent field from the triangulation tri.

DelaunayTriangulation.get_adjacent2vertex — Method

get_adjacent2vertex(tri::Triangulation)

Returns the adjacent2vertex field from the triangulation tri.

DelaunayTriangulation.get_graph — Method

get_graph(tri::Triangulation)

Returns the graph field from the triangulation tri.

DelaunayTriangulation.get_boundary_nodes — Method

get_boundary_nodes(tri::Triangulation)

Returns the boundary_nodes field from the triangulation tri.

DelaunayTriangulation.get_boundary_edge_map — Method

get_boundary_edge_map(tri::Triangulation)

Returns the boundaryedgemap field from the triangulation tri.

DelaunayTriangulation.get_boundary_map — Method

get_boundary_map(tri::Triangulation)

Returns the boundary_map field from the triangulation tri.

DelaunayTriangulation.get_boundary_index_ranges — Method

get_boundary_index_ranges(tri::Triangulation)

Returns the boundaryindexranges field from the triangulation tri.

DelaunayTriangulation.get_constrained_edges — Method

get_constrained_edges(tri::Triangulation)

Returns the constrained_edges field from the triangulation tri.

DelaunayTriangulation.get_all_constrained_edges — Method

get_all_constrained_edges(tri::Triangulation)

Returns the allconstrainededges field from the triangulation tri.

DelaunayTriangulation.get_convex_hull — Method

get_convex_hull(tri::Triangulation)

Returns the convex_hull field from the triangulation tri.

DelaunayTriangulation.get_representative_point_list — Method

get_representative_point_list(tri::Triangulation)

Returns the representativepointlist field from the triangulation tri.

There are several useful methods available for working with triangulations. We split these into several sections.

Adjacent Methods

DelaunayTriangulation.get_adjacent — Method

get_adjacent(tri, uv; check_existence=Val(has_multiple_segments(tri)))
get_adjacent(tri, u, v; check_existence=Val(has_multiple_segments(tri)))

Returns get_adjacent(tri, uv; check_existence) or get_adjacent(tri, u, v; check_existence).

DelaunayTriangulation.add_adjacent! — Method

add_adjacent!(tri::Triangulation, uv, w)
add_adjacent!(tri::Triangulation, u, v, w)

Calls add_adjacent!(get_adjacent(tri), uv, w) or add_adjacent!(get_adjacent(tri), u, v, w), adding the edge uv = (u, v) to the adjacent map of tri with corresponding vertex w.

DelaunayTriangulation.delete_adjacent! — Method

delete_adjacent!(tri::Triangulation, uv)
delete_adjacent!(tri::Triangulation, u, v)

Calls delete_adjacent!(get_adjacent(tri), uv) or delete_adjacent!(get_adjacent(tri), u, v), removing the edge uv = (u, v) from the adjacent map.

Adjacent2Vertex Methods

DelaunayTriangulation.get_adjacent2vertex — Method

get_adjacent2vertex(tri::Triangulation, w)

Returns get_adjacent2vertex(get_adjacent2vertex(tri), w), the edges (u, v) in tri that are adjacent to w.

DelaunayTriangulation.add_adjacent2vertex! — Method

add_adjacent2vertex!(tri::Triangulation, w, uv)
add_adjacent2vertex!(tri::Triangulation, w, u, v)

Calls add_adjacent2vertex!(get_adjacent2vertex(tri), w, uv) or add_adjacent2vertex!(get_adjacent2vertex(tri), w, u, v), pushing the edge uv = (u, v) into get_adjacent2vertex(tri, w).

DelaunayTriangulation.delete_adjacent2vertex! — Method

delete_adjacent2vertex!(tri::Triangulation, w, uv)
delete_adjacent2vertex!(tri::Triangulation, w, u, v)

Calls delete_adjacent2vertex!(get_adjacent2vertex(tri), w, uv) or delete_adjacent2vertex!(get_adjacent2vertex(tri), w, u, v), deleting the edge uv = (u, v) from get_adjacent2vertex(tri, w).

DelaunayTriangulation.delete_adjacent2vertex! — Method

delete_adjacent2vertex!(tri::Triangulation, w)

Calls delete_adjacent2vertex!(get_adjacent2vertex(tri), w), deleting the vertex w from get_adjacent2vertex(tri).

Boundary Nodes Methods

DelaunayTriangulation.has_multiple_curves — Method

has_multiple_curves(tri::Triangulation)

Returns has_multiple_curves(get_boundary_nodes(tri)), testing if tri's boundary is comprised of multiple curves.

DelaunayTriangulation.has_multiple_segments — Method

has_multiple_segments(tri::Triangulation)

Returns has_multiple_segments(get_boundary_nodes(tri)), testing if tri's boundary is comprised of multiple segments.

DelaunayTriangulation.num_curves — Method

num_curves(tri::Triangulation)

Returns num_curves(get_boundary_nodes(tri)), the number of curves in tri's boundary.

DelaunayTriangulation.num_segments — Method

num_segments(tri::Triangulation)

Returns num_segments(get_boundary_nodes(tri)), the number of segments in tri's boundary.

DelaunayTriangulation.get_boundary_nodes — Method

get_boundary_nodes(bn, mnℓ...)

Get the boundary nodes from bn corresponding to the specified indices. See getboundarynodes.

get_boundary_nodes(tri::Triangulation, mnℓ...)

Returns get_boundary_nodes(get_boundary_nodes(tri), mnℓ...).

DelaunayTriangulation.map_boundary_index — Method

map_boundary_index(tri::Triangulation, i)

Returns map_boundary_index(get_boundary_map(tri), i), mapping the boundary index i to the corresponding index in the boundary nodes.

DelaunayTriangulation.get_curve_index — Method

get_curve_index(tri::Triangulation, i)

Returns get_curve_index(get_boundary_map(tri), i), the curve corresponding to the boundary index i.

DelaunayTriangulation.get_segment_index — Method

get_segment_index(tri::Triangulation, i)

Returns get_segment_index(get_boundary_map(tri), i), the segment corresponding to the boundary index i.

DelaunayTriangulation.num_outer_boundary_segments — Method

num_outer_boundary_segments(tri::Triangulation)

Returns num_outer_boundary_segments(get_boundary_nodes(tri)), returns the number of boundary segments on the outermost boundary.

DelaunayTriangulation.get_right_boundary_node — Method

get_right_boundary_node(tri::Triangulation, k, boundary_index)

Returns get_right_boundary_node(get_adjacent(tri), k, boundary_index, get_boundary_index_ranges(tri), Val(has_multiple_segments(tri))), the node to the right of the boundary node k. It is assumed that k is on the part of the boundary of tri with index boundary_index.

DelaunayTriangulation.get_left_boundary_node — Method

get_left_boundary_node(tri::Triangulation, k, boundary_index)

Returns get_left_boundary_node(get_adjacent(tri), k, boundary_index, get_boundary_index_ranges(tri), Val(has_multiple_segments(tri))), the node to the left of the boundary node k. It is assumed that k is on the part of the boundary of tri with index boundary_index.

DelaunayTriangulation.get_boundary_index_range — Method

get_boundary_index_range(tri::Triangulation, i)

Returns map_boundary_index(get_boundary_index_ranges(tri), i), the list of boundary indices belonging to the curve corresponding to the boundary index i.

DelaunayTriangulation.get_boundary_edge_map — Method

get_boundary_edge_map(tri::Triangulation, ij)
get_boundary_edge_map(tri::Triangulation, i, j)

Returns get_boundary_edge_map(get_boundary_nodes(tri), ij) or get_boundary_edge_map(get_boundary_nodes(tri), construct_edge(edge_type(tri), i, j)), returning a Tuple pos such that get_boundary_nodes(get_boundary_nodes(tri, pos[1]), pos[2]) is the edge ij = (i, j).

DelaunayTriangulation.insert_boundary_node! — Method

insert_boundary_node!(tri::Triangulation, pos, node)

Calls insert_boundary_node!(get_boundary_nodes(tri), pos, node), splitting the boundary node at pos at node.

DelaunayTriangulation.split_boundary_edge! — Method

split_boundary_edge!(tri::Triangulation, edge, node)
split_boundary_edge!(tri::Triangulation, i, j, node)

Given a triangulation tri, an edge = (i, j), and a node on edge, this function splits the boundary edge into two edges edge = (i, node) and (node, j), updating tri.boundary_nodes and tri.boundary_edge_map accordingly.

See also split_boundary_edge_at_collinear_segments!.

DelaunayTriangulation.split_boundary_edge_at_collinear_segments! — Method

split_boundary_edge_at_collinear_segments!(tri::Triangulation, collinear_segments)

Given a triangulation tri and a list of collinear segments collinear_segments, assumed to represent a boundary edge (initial(first(collinear_segments)), terminal(last(collinear_segments))), splits the edge at the collinear_segments and updates tri.boundary_nodes and tri.boundary_edge_map accordingly.

See also split_boundary_edge!.

DelaunayTriangulation.contains_boundary_edge — Method

contains_boundary_edge(tri::Triangulation, e)
contains_boundary_edge(tri::Triangulation, i, j)

Tests if the triangulation tri has the constrained boundary edge e = (i, j), returning e ∈ keys(get_boundary_edge_map(tri)).

DelaunayTriangulation.merge_constrained_edges — Method

merge_constrained_edges(bn_map, boundary_nodes, constrained_edges::Es)
merge_constrained_edges(tri::Triangulation, bn_map=get_boundary_map(tri))

Merges the boundary edges, defined via boundary_nodes or get_boundary_nodes(tri), and the constrained_edges into a single collection. bn_map is used to iterate over all the boundary edges.

DelaunayTriangulation.get_all_boundary_nodes — Method

get_all_boundary_nodes(tri::Triangulation)

Returns a Set of all the boundary nodes in the triangulation tri.

DelaunayTriangulation.all_boundary_indices — Method

all_boundary_indices(tri::Triangulation)

Returns all the boundary indices in the triangulation tri.

DelaunayTriangulation.delete_boundary_node! — Method

delete_boundary_node!(tri::Triangulation, pos)

Calls delete_boundary_node!(get_boundary_nodes(tri), pos), deleting the boundary node at pos.

Missing docstring.

Missing docstring for merge_boundary_node!(::Triangulation, ::Any, ::Any). Check Documenter's build log for details.

Convex Hull Methods

DelaunayTriangulation.get_convex_hull_indices — Method

get_convex_hull_indices(tri::Triangulation)

Calls get_indices(get_convex_hull(tri)).

DelaunayTriangulation.convex_hull! — Method

convex_hull!(tri::Triangulation; reconstruct=is_constrained(tri))

Computes the convex hull of the points included in the triangulation tri. If reconstruct, then the method in convex_hull will be used, whereas if !reconstruct then the triangulation's boundary is extracted to get the convex hull directly. Note that this latter method fails if there are any constrained edges on the boundary of tri.

Edges Methods

DelaunayTriangulation.edge_type — Method

edge_type(tri::Triangulation)

Returns the type used for representing edges in tri.

edge_type(vor::VoronoiTessellation)

Returns the type of the edges in the Voronoi tessellation vor.

DelaunayTriangulation.num_edges — Method

num_edges(tri::Triangulation)

Returns num_edges(get_graph(tri)), the number of edges in tri.

DelaunayTriangulation.each_edge — Method

each_edge(tri::Triangulation)

Returns get_edges(get_graph(tri)). This iterator over the edges could include ghost edges.

See also each_solid_edge and each_ghost_edge

DelaunayTriangulation.each_constrained_edge — Method

each_constrained_edge(tri::Triangulation)

Returns each_edge(get_all_constrained_edges(tri)), an iterator over all constrained_edges (including all constrained boundary edges)`.

DelaunayTriangulation.contains_constrained_edge — Method

contains_constrained_edge(tri::Triangulation, e)
contains_constrained_edge(tri::Triangulation, i, j)

Tests if tri contains the constrained edge e = (i, j) (or reverse_edge(e) = (j, i)).

DelaunayTriangulation.num_ghost_edges — Method

num_ghost_edges(tri::Triangulation)

Returns the number of ghost edges in tri.

DelaunayTriangulation.num_solid_edges — Method

num_solid_edges(tri::Triangulation)

Returns the number of solid edges in tri.

DelaunayTriangulation.each_solid_edge — Function

each_solid_edge(tri)

Returns an iterator over all solid edges of the triangulation tri, i.e. over all edges that are not ghost edges.

See also each_edge and each_ghost_edge.

DelaunayTriangulation.each_ghost_edge — Function

each_ghost_edge(tri)

Returns an iterator over all ghost edges of the triangulation tri, i.e. over all edges that have a boundary index as one of its indices.

See also each_edge and each_solid_edge.

DelaunayTriangulation.sort_edge_by_degree — Method

sort_edge_by_degree(tri::Triangulation, e)

Returns sort_edge_by_degree(e, get_graph(tri)).

DelaunayTriangulation.split_constrained_edge! — Method

split_constrained_edge!(tri::Triangulation, constrained_edge, collinear_segments)

Calls split_constrained_edge!(get_constrained_edges(tri), constrained_edge, collinear_segments), splitting the constrained_edge which is assumed to represent the union of the collinear_segments so that we instead store those collinear_segments.

Graph Methods

DelaunayTriangulation.get_edges — Method

get_graph(tri::Triangulation)

Returns get_edges(get_graph(tri)).

DelaunayTriangulation.get_vertices — Method

get_vertices(tri::Triangulation)

Returns get_vertices(get_graph(tri)), giving all vertices present in the triangulation. This will include ghost vertices.

See also each_vertex, each_solid_vertex and each_ghost_vertex.

DelaunayTriangulation.get_neighbours — Method

get_neighbours(tri::Triangulation)

Returns get_neighbours(get_graph(tri)), the set of neighbourhoods in the triangulation tri.

DelaunayTriangulation.get_neighbours — Method

get_neighbours(tri::Triangulation, u)

Returns get_neighbours(get_graph(tri), u), the set of neighbours of u.

DelaunayTriangulation.add_vertex! — Method

add_vertex!(tri::Triangulation, u...)

Calls add_vertex!(get_graph(tri), u...), adding the vertex u into tri.

DelaunayTriangulation.num_neighbours — Method

num_neighbours(tri::Triangulation, u)

Returns num_neighbours(get_graph(tri), u), the number of neighbours of u.

DelaunayTriangulation.add_neighbour! — Method

add_neighbour!(tri::Triangulation, u, v...)

Calls add_neighbour!(get_graph(tri), u, v...), adding each v into the neighbourhood of each u.

DelaunayTriangulation.delete_neighbour! — Method

delete_neighbour!(tri::Triangulation, u, v...)

Calls delete_neighbour!(get_graph(tri), u, v...), deleting each v from the neighbourhood of u.

DelaunayTriangulation.delete_vertex! — Method

delete_vertex!(tri::Triangulation, u...)

Calls delete_vertex!(get_graph(tri), u...), deleting each vertex u from tri.

DelaunayTriangulation.delete_boundary_vertices_from_graph! — Method

delete_boundary_vertices_from_graph!(tri::Triangulation)

Calls delete_boundary_vertices_from_graph!(get_graph(tri)), deleting all ghost vertices from the graph.

DelaunayTriangulation.each_vertex — Method

each_vertex(tri::Triangulation)

Returns each_vertex(get_graph(tri)), an iterator over all vertices present in the triangulation.

Warning

This iterator will include ghost vertices. If you want to exclude these, see each_solid_vertex. Alternatively, if you only want ghost vertices, see each_ghost_vertex.

DelaunayTriangulation.num_vertices — Method

num_vertices(tri::Triangulation)

Returns num_vertices(get_graph(tri)), the number of vertices in the triangulation.

Points Methods

DelaunayTriangulation.get_point — Method

get_point(tri::Triangulation, i...)

Returns the ith point of `tri. Boundary indices are automatically handled.

DelaunayTriangulation.each_point_index — Method

each_point_index(tri::Triangulation)

Returns an iterator over the indices of the points of tri.

See also each_vertex, `eachsolidvertex and each_ghost_vertex.

DelaunayTriangulation.each_point — Method

each_point(tri::Triangulation)

Returns an iterator over the points of tri.

See also each_vertex, `eachsolidvertex and each_ghost_vertex.

DelaunayTriangulation.num_points — Method

num_points(tri::Triangulation)

Returns the number of points of tri.

Note

Note that this is just the size of get_points(tri), but if there are some missing points then this will not match the number in the triangulation itself. Use num_vertices to get the number of vertices in the triangulation.

DelaunayTriangulation.push_point! — Method

push_point!(tri::Triangulation, x, y) = push_point!(get_points(tri), x, y)
push_point!(tri::Triangulation, p) = push_point!(get_points(tri), p)

Pushes the point p = (x, y) into get_points(tri0.

Note

This does not add the point p into the triangulation itself. See add_point! for this.

DelaunayTriangulation.each_solid_vertex — Function

each_solid_vertex(tri::Triangulation)

Returns an iterator over the solid vertices in the triangulation tri, i.e. those that are not ghost vertices.

DelaunayTriangulation.each_ghost_vertex — Function

each_ghost_vertex(tri::Triangulation)

Returns an iterator over the ghost vertices in the triangulation tri.

DelaunayTriangulation.num_solid_vertices — Function

num_solid_vertices(tri::Triangulation)

Returns the number of solid vertices of tri.

num_solid_vertices(stats::TriangulationStatistics)

Returns the numsolidvertices field from the triangulation statistics stats.

DelaunayTriangulation.num_ghost_vertices — Function

num_ghost_vertices(tri::Trianngulation)

Returns the number of ghost vertices of tri.

num_ghost_vertices(stats::TriangulationStatistics)

Returns the numghostvertices field from the triangulation statistics stats.

Triangles Methods

DelaunayTriangulation.triangle_type — Method

triangle_type(tri::Triangulation)

Returns the type used for representing triangles in tri.

triangle_type(vor::VoronoiTessellation)

Returns the type of the triangles used in the triangulation underlying the Voronoi tessellation vor.

DelaunayTriangulation.num_triangles — Method

num_triangles(tri::Triangulation) = num_triangles(get_triangles(tri))

Returns the number of triangles in tri, including ghost triangles.

DelaunayTriangulation.each_triangle — Method

each_triangle(tri::Triangulation)

Returns an iterator over each triangle in tri. This iterator could include ghost triangles.

See also each_solid_triangle and each_ghost_triangle.

DelaunayTriangulation.contains_triangle — Method

contains_triangle(tri::Triangulation, T)
contains_triangle(tri::Triangulation, i, j, k)

Tests if tri contains the triangle T.

DelaunayTriangulation.construct_positively_oriented_triangle — Method

construct_positively_oriented_triangle(tri::Triangulation, i, j, k)

Constructs a triangle T with vertices (i, j, k) in tri such that T is positively oriented.

DelaunayTriangulation.num_ghost_triangles — Method

num_ghost_triangles(tri::Triangulation)

Returns the number of ghost triangles in tri.

DelaunayTriangulation.num_solid_triangles — Method

num_solid_triangles(tri::Triangulation)

Returns the number of solid triangles in tri.

DelaunayTriangulation.each_solid_triangle — Function

each_solid_triangle(tri)

Returns an iterator over all triangles of tri that are not ghost triangles.

DelaunayTriangulation.each_ghost_triangle — Function

each_ghost_triangle(tri)

Returns an iterator over all triangles of tri that are ghost triangles.

Predicates Methods

DelaunayTriangulation.is_boundary_edge — Method

is_boundary_edge(tri::Triangulation, ij)
is_boundary_edge(tri::Triangulation, i, j)

Returns is_boundary_edge(ij, get_adjacent(tri)) or is_boundary_edge(i, j, get_adjacent(tri)), respectively, testing if ij = (i, j) belongs to the boundary of the triangulation tri, i.e. (i, j) adjoins a ghost vertex.

DelaunayTriangulation.is_boundary_triangle — Method

is_boundary_triangle(tri::Triangulation, i, j, k)
is_boundary_triangle(tri::Triangulation, T)

Returns is_boundary_triangle(i, j, k, get_adjacent(tri)) or is_boundary_triangle(T, get_adjacent(tri)), testing if at least one of the edges (u, v) of T = (i, j, k) satisfies is_boundary_edge(tri, u, v).

DelaunayTriangulation.triangle_orientation — Method

triangle_orientation(tri::Triangulation, i, j, k)
triangle_orientation(tri::Triangulation, T)

Computes the orientation of the triangle T = (i, j, k) in tri.

DelaunayTriangulation.point_position_relative_to_circumcircle — Method

point_position_relative_to_circumcircle(tri::Triangulation, i, j, k, ℓ)
point_position_relative_to_circumcircle(tri::Triangulation, T, ℓ)

Computes the position of the ℓth point of tri relative to the circumcircle of the triangle T = (i, j, k) in tri.

DelaunayTriangulation.point_position_relative_to_line — Method

point_position_relative_to_line(tri::Triangulation, i, j, u)

Computes the position of the uth point of tri relative to the line segment with indices (i, j) in tri.

DelaunayTriangulation.point_closest_to_line — Method

point_closest_to_line(tri::Triangulation, i, j, u, v)

Tests which of the points u or v is closest to the line segment with indices (i, j) in tri.

DelaunayTriangulation.point_position_on_line_segment — Method

point_position_on_line_segment(tri::Triangulation, i, j, u)

Given vertices i, j, and u that are collinear, computes the position of u on the line segment (i, j).

DelaunayTriangulation.line_segment_intersection_type — Method

line_segment_intersection_type(tri::Triangulation, i, j, u, v)

Given two lines with indices (i, j) and (u, v), respectively, in tri, computes the number of intersections.

DelaunayTriangulation.point_position_relative_to_triangle — Method

point_position_relative_to_triangle(tri::Triangulation, i, j, k, u)
point_position_relative_to_triangle(tri::Triangulation, T, u)

Computes the position of the uth point of tri relative to the triangle T = (i, j, k) in tri.

DelaunayTriangulation.triangle_line_segment_intersection — Method

triangle_line_segment_intersection(tri::Triangulation, i, j, k, u, v)
triangle_line_segment_intersection(tri::Triangulation, T, e)

Computes the type of the intersection of the triangle T = (i, j, k) in tri with the line segment e = (u, v).

DelaunayTriangulation.is_outer_boundary_index — Method

is_outer_boundary_index(tri::Triangulation, i)

Returns true if is_boundary_index(i) and this boundary index i corresponds to the outermost boundary of tri. Returns false otherwise.

DelaunayTriangulation.is_outer_boundary_node — Method

is_outer_boundary_node(tri::Triangulation, i)

Returns true if i is a boundary node belonging to the outermost boundary of tri. Returns false otherwise.

DelaunayTriangulation.is_boundary_node — Method

is_boundary_node(tri::Triangulation, i)

Returns true if i is a node belonging to any boundary of tri.

DelaunayTriangulation.edge_exists — Method

edge_exists(tri::Triangulation, i, j)
edge_exists(tri::Triangulation, ij)

Returns true if the edge (i, j) exists in tri, and false otherwise.

DelaunayTriangulation.has_ghost_triangles — Method

has_ghost_triangles(tri::Triangulation)

Returns true if tri has ghost triangles, and false otherwise.

DelaunayTriangulation.has_boundary_nodes — Method

has_boundary_nodes(tri::Triangulation)

Returns true if tri has boundary nodes, and false otherwise.

DelaunayTriangulation.is_legal — Method

is_legal(tri::Triangulation, i, j)

Returns true if the edge (i, j) is legal in tri, and false otherwise. We also define constrained edges to be legal, as are ghost edges.

DelaunayTriangulation.find_edge — Method

find_edge(tri::Triangulation, T, ℓ)

Given a point ℓ that is on an edge of the triangle T in tri, returns the edge that ℓ is on.

DelaunayTriangulation.is_constrained — Method

is_constrained(tri::Triangulation, i, j)

Returns true if tri has any constrained edges, and false otherwise.

Representative Points Methods

DelaunayTriangulation.compute_representative_points! — Method

compute_representative_points!(tri::Triangulation; use_convex_hull=!has_multiple_segments(tri) && num_boundary_edges(get_boundary_nodes(tri)) == 0)

Updates get_representative_point_list(tri) to match the current position of the boundaries. If there are no boundary nodes, use_convex_hull instead represents them using the indices of the convex hull.

DelaunayTriangulation.get_representative_point_coordinates — Method

get_representative_point_coordinates(tri::Triangulation, i)

Returns the coordinates of the ith representative point in tri.

DelaunayTriangulation.reset_representative_points! — Method

reset_representative_points!(tri::Triangulation)

Resets get_representative_point_list(tri).

DelaunayTriangulation.update_centroid_after_addition! — Method

update_centroid_after_addition!(tri::Triangulation, i, p)

After the point p has been added into tri, this updates the ith representative point in get_representative_point_list(tri), treating it as if it were a centroid.

DelaunayTriangulation.update_centroid_after_deletion! — Method

update_centroid_after_deletion!(tri::Triangulation, i, p)

After the point p has been deleted from tri, this updates the ith representative point in get_representative_point_list(tri), treating it as if it were a centroid.

DelaunayTriangulation.new_representative_point! — Method

new_representative_point!(tri::Triangulation, i)

Adds a new representative point to get_representative_point_list(tri) with index i.