Interface

The package makes its simple for customing the interface used for defining points, edges, and triangles, as described in the docstring for Interfaces (see the end of this section). Without any customisation, the default forms are:

Edges: NTuple{2, Int64}.

Collections of edges: Set{NTuple{2, Int64}}.

Triangles: NTuple{3, Int64}.

Collections of triangles: Set{NTuple{3, Int64}}

We also give support for customing how points are represented, and by default we support collections of points given as matrices (with each point its own column), or vectors of vectors. The number type used for representing coordinates has to be Float64 to support ExactPredicates.jl, although if you like you could customise orient_predicate and incircle_predicate, even circumventing ExactPredicates.jl if you like. See the predicates section in the sidebar for a further discussion of changing these predicate definitions.

We also provide a customisable interface for representing boundary nodes, although for unconstrained triangulations this is relevant. By default, we support boundary nodes represented according to the following, where we let BN refer to the collection of boundary nodes:

Vector{Int64}: In this case, there is only one fixed boundary and it is represented as a contiguous set of nodes. We must have BN[begin] == BN[end], and the nodes must be listed in counter-clockwise order.

Vector{Vector{Int64}}: In this case, there is only one fixed boundary, but it is made up of separate segments, with BN[n] the nodes for the nth segment. This makes it possible to more easily support, for example, a domain with different boundary conditions on separate parts of the boundary. We must have BN[n][end] == BN[n+1][begin] and BN[end][end] == BN[begin][begin], and each segment must be listed in counter-clockwise order.

Vector{Vector{Vector{Int64}}}: In this case, there are multiple fixed boundaries, each of which are assumed to be made up of separate segments. This makes it possible to support multiply-connected domains, e.g. an annulus with each circle split into its lower and upper halves. Here, BN[m][n] is the set of nodes for the nth segment of the mth boundary curve, and BN[begin] the outer-most boundary curve and BN[m], m > 1, nodes for curves contained within BN[begin]. As in the previous case, BN[m][n][end] == BN[m][n+1][end] and BN[m][end][end] == BN[m][begin][begin] for each m. Moreover, BN[begin] should be a counter-clockwise list of nodes while BN[m] is a clockwise list of nodes for m > 1.

For more information about how we handle boundaries, and how they are handled in our data structures, see the boundary handling section in the sidebar.

The complete docstring that we use to also describe these interfaces is given below. (Accessible via ?Interfaces). See the next subsections for further detail.

DelaunayTriangulation.Interfaces — Function

Interfaces

The main objects used throughout this package, namely points, edges, triangles, and boundaries, have interfaces that are fully customisable. We describe these interfaces below, along with all the methods that need to be defined.

For these interfaces, we define default methods that support Tuples, AbstractVectors, etc., so you do not need to extend any of these for normal use. Moreover, while we do allow for different index types to be using, all still have to subtypes of Integer (for now?).

Triangle Interface

Here we describe the interface used for describing triangles and collections of triangles.

Individual Triangles

A triangle is assumed to be of the form (i, j, k), but the object used for storing these three indices can be customised. The following methods are used for working with triangles; see their docstrings for more information.

construct_triangle: Must be defined.

geti: Must be defined.

getj: Must be defined.

getk: Must be defined.

indices: Uses the geti, getj, and getk definitions.

integer_type: Must be defined.

triangle_edges: Uses the geti, getj, and getk definitions.

rotate_triangle: Uses the geti, getj, and getk definitions.

construct_positively_oriented_triangle: Uses existing definitions.

compare_triangles: Uses the geti, getj, and getk definitions.

sort_triangle: Uses the indices and construct_triangle definitions.

Collection of Triangles

A collection of triangles simply stores many triangles. The following methods are used for working with these collections; see their docstrings for more information. Note that these collections must be mutable.

initialise_triangles: Must be defined.

triangle_type: Must be defined.

num_triangles: Must be defined.

contains_triangle: This calls Base.in and contains_triangle.

add_to_triangles!: Must be defined.

add_triangle!: This simply calls add_to_triangles!.

delete_from_triangles!: Must be defined.

delete_triangle!: This simply calls delete_from_triangles!.

each_triangle: Must be defined.

Base.in: Must be defined.

Base.sizehint!: Must be defined.

Base.unique!: Must be defined, unless your collection is a Set.

compare_triangle_collections: Calls num_triangles, each_triangle, and contains_triangle.

sort_triangles: Uses the initialise_triangles, each_triangle, sort_triangle, and add_triangle! definitions.

remove_duplicate_triangles: Uses Base.unique!.

Note that Triangulations also define each_solid_triangle and each_ghost_triangle.

Edge Interface

Here we describe the interface used for describing edges and collections of edges.

Individual Edges

An edge is assumed to be of the form (i, j), but the object used for storing these two indices can be customised. The following methods are used for working with edges; see their docstrings for more information.

construct_edge: Must be defined.

initial: Must be defined.

terminal: Must be defined.

edge_indices: Uses the initial and terminal definitions.

Collection of Edges

A collection of edges simply stores many edges. The following methods are used for working with these collections; see their docstrings for more information. Note that these collections must be mutable.

initialise_edges: Must be defined.

edge_type: Must be defined.

num_edges: Must be defined.

contains_edge: Must be defined.

add_to_edges!: Must be defined.

add_edge!: This simply calls add_to_edges!.

delete_from_edges!: Must be defined.

delete_edge!: This simply calls delete_from_edges!.

each_edge: Must be defined.

random_edge: Uses the each_edge definition.

is_empty: Simply uses the isempty definition from Base.

Point Interface

Here we describe the interface used for describing points and collections of points.

Individual Points

A point is assumed to be of the form (x, y), but the object used for storing these two coordinates can be customised. The following methods are used for working with points; see their docstrings for more information.

getx: Must be defined.

gety: Must be defined.

getxy: Uses the getx and gety definitions.

Collection of Points

A collection of points simply stores many points. The following methods are used for working with these collections; see their docstrings for more information.

getpoint: Must be defined.

get_point: This simply calls getpoint.

each_point_index: Must be defined.

each_point: Must be defined.

num_points: Must be defined.

points_are_unique: Makes use of the existing methods.

lexicographic_order: Makes use of the existing methods.

number_type: Must be defined.

Boundary Nodes Interface

Here we describe the interface used for representing the boundaries. As described in Triangulation, this interface can be used for representing either a collection of curves each made up of multiple segments, a collection of segments, or a single continuous curve. The following functions facilitate these possibilities; see the corresponding docstrings for more information.

has_multiple_curves: Must be defined.

has_multiple_segments: Must be defined.

num_curves: Must be defined.

num_segments: Must be defined.

num_boundary_edges: Must be defined.

getboundarynodes: Must be defined.

get_boundary_nodes: This just calls getboundarynodes.

each_boundary_node: Must be defined.

construct_boundary_map: Makes use of the existing methods.

construct_boundary_index_ranges: Makes use of the existing methods.

map_boundary_index: Makes use of the result from construct_boundary_map.

get_curve_index: Makes use of the result from construct_boundary_map.

get_segment_index: Makes use of the result from construct_boundary_map.

num_outer_boundary_segments: Makes use of the existing methods.