Triangulation

The most important structure of the package is the Triangulation data structure. Its complete definition is given below, and then afterwards we give all the docstrings for methods that are useful for working with triangulations.

DelaunayTriangulation.Triangulation — Type

Triangulation{P,Ts,I,E,Es,BN,B,BIR}

Struct for a triangulation.

See also triangulate and generate_mesh.

Fields

points::P

The nodes in the triangulation.

triangles::Ts

The triangles in the triangulation. All triangles are positively oriented.

adjacent::Adjacent{I, E}

The Adjacent map, mapping edges to vertices that together form a positively oriented triangle in triangles. See also Adjacent.

adjacent2vertex::Adjacent2Vertex{I,Es,E}

The Adjacent2Vertex map, mapping vertices to all edges that together form a positively oriented triangle in triangles. See also Adjacent2Vertex.

graph::Graph{I}

The Graph, mapping vertices to all other vertices that are connected with that vertex via an edge of a triangle in triangles. See also Graph.

boundary_nodes::BN

The boundary nodes, with the outer boundary given in counter-clockwise order and the inner boundaries given in clockwise order. The default (default because you could customise it if you wish, see Interfaces) form of this field depends on your domain:

– Vector{Int64}

If you provide no boundary curves or provide a single connected boundary curve, this will be a counter-clockwise vector of boundary nodes with boundary_nodes[begin] == boundary_nodes[end]. Note that if no boundary curve was provided, this gives the convex hull of points.

– Vector{Vector{Int64}}

In this case, you have provided a single boundary curve but separated into different parts. In this case, boundary_nodes[n] gives the boundary nodes for the nth segment. Additionally, boundary_nodes[n][end] == boundary_nodes[n+1][begin] and boundary_nodes[end][end] == boundary_nodes[begin][begin]. The nodes should be counter-clockwise in this case also.

– Vector{Vector{Vector{Int64}}}

In this case, you have provided multiple boundary curves each separated into different parts, e.g. an annulus with each circle split into its lower and upper halves. The component boundary_nodes[m][n] gives the boundary nodes for the nth segment of the mth boundary curve, with boundary_nodes[begin] the outer-most boundary and boundary_nodes[i] the inner boundaries (i > 1). As in the previous case, boundary_nodes[m][n][end] == boundary_nodes[m][n+1][begin] and boundary_nodes[m][end][end] == boundary_nodes[m][begin][begin]. With this form, boundary_nodes[m] should be a counter-clockwise list of nodes for m == 1, while for m > 1 it` should be a clockwise list of nodes.

boundary_map::B

This is an OrderedDict that maps a given boundary index to its position in boundary_nodes. For example, if boundary_map[-4] = (2, 3), this means that the boundary index -4 corresponds to the nodes in get_boundary_nodes(boundary_nodes, 2, 3). If there is just a single continuous curve, so that boundary_nodes acts like a vector of integers, then boundary_map[-1] simply returns boundary_nodes. See also construct_boundary_map and map_boundary_index. The ordering is so that the lowest magnitude boundary index comes first.

boundary_index_ranges::BIR

This is an OrderedDict that maps a boundary index to a range of all other boundary indices that the corresponding boundary curve could correspond to. For example, for a curve with four segments, there are four possible boundary indices that a segment could correspond to - this Dict will extract this range. See also construct_boundary_index_ranges.

constrained_edges::Es

This is the set of extra edges added into the triangulation that you have provided. Note that this will not included any of the other constrained edges from boundary_nodes.

convex_hull::ConvexHull{P,Vector{I}}

This will be a vector of integers corresponding to indices in the points that together give the convex hull of the set of points, with convex_hull[begin] == convex_hull[end].

Constructors

There are several ways to construct this struct directly (though you should probably be using triangulate or generate_mesh).

Default Constructor

The default constructor is available, i.e.

Triangulation(points, triangles, adjacent, adjacent2vertex, graph, boundary_nodes, constrained_edges, boundary_map, convex_hull)

Empty Triangulation

If you want to initialise the triangulation, you can use

Triangulation(points::P;
    IntegerType::Type{I}=Int64,
    EdgeType::Type{E}=NTuple{2,IntegerType},
    TriangleType::Type{V}=NTuple{3,IntegerType},
    EdgesType::Type{Es}=Set{EdgeType},
    TrianglesType::Type{Ts}=Set{TriangleType},
    boundary_nodes::BN=IntegerType[],
    constrained_edges=initialise_edges(EdgesType)) where {P,Ts,I,E,Es,BN,V}

Existing Triangulation Objects

The second constructor, mainly existing so that generate_mesh can convert its results into a Triangulation, takes in existing triangles, points, and boundary_nodes and creates a Triangulation:

Triangulation(points::P, triangles::T, boundary_nodes::BN;
    IntegerType::Type{I}=Int64,
    EdgeType::Type{E}=NTuple{2,IntegerType},
    TriangleType::Type{V}=NTuple{3,IntegerType},
    EdgesType::Type{Es}=Set{EdgeType},
    TrianglesType::Type{Ts}=Set{TriangleType},
    add_ghost_triangles=false) where {I, E, V, Es, Ts, T}

You can set add_ghost_triangles = true to add ghost triangles into the structure.

Extended help

There are many functions available for working with a Triangulation, some of which we list below.

Accessors:

get_points

get_triangles

get_adjacent

get_adjacent2vertex

get_graph

get_boundary_nodes

get_boundary_map

get_boundary_index_ranges

get_constrained_edges

get_convex_hull

Operations:

add_triangle!

delete_triangle!

add_ghost_triangles!

delete_ghost_triangles!

add_boundary_information!

Point location:

brute_force_search

jump_and_march

Working with the Adjacent field:

get_adjacent

add_adjacent!

delete_adjacent!

Working with the Adjacent2Vertex field:

get_adjacent2vertex

add_adjacent2vertex!

delete_adjacent2vertex!

Working with the Graph field:

get_edges

get_vertices

get_neighbours

add_vertex!

add_neighbour!

delete_neighbour!

delete_vertex!

delete_boundary_vertices_from_graph!

Working with the boundary_nodes field:

has_multiple_curves

has_multiple_segments

num_curves

num_segments

get_boundary_nodes

map_boundary_index

get_curve_index

get_segment_index

num_outer_boundary_segments

get_right_boundary_node

get_left_boundary_node

Working with the convex_hull field:

get_convex_hull_indices

convex_hull!

Working with the triangles field:

triangle_type

num_triangles

each_triangle

each_solid_triangle

each_ghost_triangle

contains_triangle

Working with edges:

edge_type

num_edges

each_edge

Working with the points field:

get_point

each_point_index

each_point

num_points

Predicates:

is_boundary_edge

is_boundary_triangle

triangle_orientation

point_position_relative_to_circumcircle

point_position_relative_to_line

point_position_on_line_segment

line_segment_intersection_type

point_position_relative_to_triangle

is_outer_boundary_index

is_outer_boundary_node

edge_exists

has_ghost_triangles

has_boundary_nodes

Miscellaneous:

integer_type

number_type

compute_representative_points!

clear_empty_features!

find_edge

is_constrained

all_boundary_indices

Useful Methods

DelaunayTriangulation.get_adjacent — Method

get_adjacent(tri::Triangulation)

Returns the adjacent field from the triangulation tri.

DelaunayTriangulation.get_adjacent — Method

get_adjacent(tri::Triangulation, uv; check_existence::V=Val(false)) where {V}
get_adjacent(tri::Triangulation, u, v; check_existence::V=Val(false)) where {V}

Given the edge (u, v), gets the vertex w such that (u, v, w) is a positively oriented triangle in the triangulation tri.

Use check_existence=Val(true) to be safe against ghost edges corresponding to neighbouring boundary indices.

DelaunayTriangulation.add_adjacent! — Method

add_adjacent!(tri::Triangulation, uv, w)
add_adjacent!(tri::Triangulation, u, v, w)

Given the edge (u, v), adds the vertex w into the adjacent map of the triangulation tri, so that (u, v, w) is a positively oriented triangle in tri.

DelaunayTriangulation.delete_adjacent! — Method

delete_adjacent!(tri::Triangulation, uv)
delete_adjacent!(tri::Triangulation, u, v)

Given the edge (u, v), deletes the key (u, v) from the adjacent map of the triangulation tri.

DelaunayTriangulation.get_adjacent2vertex — Method

get_adjacent2vertex(tri::Triangulation)

Returns the adjacent2vertex field from the triangulation tri.

DelaunayTriangulation.get_adjacent2vertex — Method

get_adjacent2vertex(tri::Triangulation, w)

Given the vertex w, returns the set of edges (u, v) from the adjacent2vertex map such that (u, v, w) is a positively oriented triangle in tri.

DelaunayTriangulation.add_adjacent2vertex! — Method

add_adjacent2vertex!(tri::Triangulation, w, uv)
add_adjacent2vertex!(tri::Triangulation, w, u, v)

Given the vertex w and an edge (u, v), pushes the edge (u, v) into the set of edges defined by the key w in the adjacent2vertex map of the triangulation tri.

DelaunayTriangulation.delete_adjacent2vertex! — Method

delete_adjacent2vertex!(tri::Triangulation, w, uv)
delete_adjacent2vertex!(tri::Triangulation, w, u, v)

Given the vertex w and an edge (u, v), deletes the edge (u, v) from the set of edges defined by the key w in the adjacent2vertex map of the triangulation tri.

DelaunayTriangulation.delete_adjacent2vertex! — Method

delete_adjacent2vertex!(tri::Triangulation, w)

Given the vertex w, deletes the set of edges (u, v) that define positively oriented triangles (u, v, w) from the adjacent2vertex map of the triangulation tri.

DelaunayTriangulation.get_graph — Method

get_graph(tri::Triangulation)

Returns the graph field from the triangulation tri.

DelaunayTriangulation.get_edges — Method

get_edges(tri::Triangulation)

Given a triangulation tri, returns the set of edges in tri. These edges will not be oriented, i.e. if (i, j) is in the set then (j, i) will not be.

DelaunayTriangulation.get_vertices — Method

get_vertices(tri::Triangulation)

Given a triangulation tri, returns the set of vertices in tri. Note that this will include any ghost vertices.

DelaunayTriangulation.get_neighbours — Method

get_neighbours(tri::Triangulation)

Given a triangulation tri, returns the set of neighbourhoods.

DelaunayTriangulation.get_neighbours — Method

get_neighbours(tri::Triangulation, u)

Given a vertex u and a triangulation tri, returns the set of vertices v such that (u, v) is an edge of the triangulation for each v in the set.

DelaunayTriangulation.add_vertex! — Method

add_vertex!(tri::Triangulation, u...)

Given vertices u... and a triangulation tri, adds all the vertices u... into the graph tri.graph.

Note that this does not insert the vertex into the triangulation - see add_point! for this.

DelaunayTriangulation.add_neighbour! — Method

add_neighbour!(tri::Triangulation, u, v...)

Given a vertex u and other vertices v..., adds all the vertices v... into the neighbourhood of u from the triangulation tri, i.e. into tri.graph.

DelaunayTriangulation.delete_neighbour! — Method

delete_neighbour!(tri::Triangulation, u, v...)

Given a vertex u and other vertices v..., deletes all the vertices v... from the neighbourhood of u from the triangulation tri, i.e. from tri.graph.

DelaunayTriangulation.delete_vertex! — Method

delete_vertex!(tri::Triangulation, u...)

Given vertices u... and a triangulation tri, deletes all the vertices u... from the graph tri.graph.

Note that this does not remove the vertex from the triangulation - see remove_point! for this.

DelaunayTriangulation.delete_boundary_vertices_from_graph! — Method

delete_boundary_vertices_from_graph!(tri::Triangulation)

Given a triangulation tri, removes all boundary vertices from the triangulation tri, i.e. all those with index less than -1 ( see is_boundary_index).

DelaunayTriangulation.get_boundary_nodes — Method

get_boundary_nodes(tri::Triangulation)

Returns the boundary_nodes field from the triangulation tri.

DelaunayTriangulation.has_multiple_curves — Method

has_multiple_curves(tri::Triangulation)

Returns true if the triangulation tri has multiple boundary curves, and false otherwise.

DelaunayTriangulation.has_multiple_segments — Method

has_multiple_segments(tri::Triangulation)

Returns true if the boundary curves of the triangulation tri are broken into multiple segments. Note that if has_multiple_curves(tri), then has_multiple_segments(tri) == true.

DelaunayTriangulation.num_curves — Method

num_curves(tri::Triangulation)

Returns the number of boundary curves of the triangulation tri.

DelaunayTriangulation.num_segments — Method

num_segments(tri::Triangulation)

Returns the number of segments of the boundary curve of the triangulation tri. This is only defined if !has_multiple_curves(tri) and has_multiple_segments(tri).

DelaunayTriangulation.get_boundary_nodes — Method

get_boundary_nodes(tri::Triangulation, mnℓ...)

Given a triangulation tri, returns the nodes corresponding to the indices in mnℓ.... For example, if tri has multiple boundary curves, then get_boundary_nodes(tri, m) returns the set of nodes for the mth curve, and get_boundary_nodes(tri, m, n) the set of nodes for the nth segment of the mth curve.

DelaunayTriangulation.map_boundary_index — Method

map_boundary_index(tri::Triangulation, i)

Given a triangulation tri, returns the position in tri.boundary_nodes corresponding to the boundary index i, i.e. returns tri.boundary_map[i].

DelaunayTriangulation.get_curve_index — Method

get_curve_index(tri::Triangulation, i)

Given a triangulation tri and a boundary index i, returns the curve number that i corresponds to. For example, if map_boundary_index(tri, i) = (m, n), the curve index would be m. If !has_multiple_curves(tri), this just returns 1.

DelaunayTriangulation.get_segment_index — Method

get_segment_index(tri::Triangulation, i)

Given a triangulation tri and a boundary index i, returns the segment number that i corresponds to. For example, if map_boundary_index(tri, i) = (m, n), the segment index would be m. If !has_multiple_segments(tri), this just returns 1.

DelaunayTriangulation.get_right_boundary_node — Method

get_right_boundary_node(tri::Triangulation, k, boundary_index)

Given a triangulation tri, a boundary node k, and a boundary_index for the associated boundary, returns the index of the point to the right of k on the outer boundary.

See also get_left_boundary_node.

DelaunayTriangulation.get_left_boundary_node — Method

get_left_boundary_node(tri::Triangulation, k)

Given a triangulation tri and an outer boundary node k, returns the index of the point to the left of k on the outer boundary.

See also get_right_boundary_node.

DelaunayTriangulation.get_convex_hull_indices — Method

get_convex_hull_indices(tri::Triangulation)

Returns indices for the ConvexHull of the triangulation tri.

DelaunayTriangulation.get_convex_hull — Method

get_convex_hull(tri::Triangulation)

Returns the convex_hull field from the triangulation tri.

DelaunayTriangulation.convex_hull! — Method

convex_hull!(tri::Triangulation; reconstruct = is_constrained(tri))

Computess the convex hull for the points in the triangulation tri, updating the field tri.convex_hull. Note that this will only be needed if you are e.g. constructing a convex hull from an existing set of points, triangles, and edges. If you already have an unconstrained Delaunay triangulation, this field will have also been updated.

If !reconstruct, the convex hull is reconstructed using convex_hull, otherwise the ghost triangles are used to construct it (if there are no ghost triangles, they are added and then removed afterwards).

See also convex_hull and ConvexHull.

DelaunayTriangulation.triangle_type — Method

triangle_type(tri::Triangulation{P,Ts}) where {P,Ts}

Given a triangulation tri, returns the type used for representing the triangles.

DelaunayTriangulation.get_triangles — Method

get_triangles(tri::Triangulation)

Returns the triangles field from the triangulation tri.

DelaunayTriangulation.num_triangles — Method

num_triangles(tri::Triangulation)

Given a triangulation tri, returns the number of triangles.

DelaunayTriangulation.each_triangle — Method

each_triangle(tri::Triangulation)

Given a triangulation tri, returns an iterator over the triangles. This iterator will return both the ghost and solid triangles (if there are no ghost triangles, then those of course are not included and each_triangle(tri) == each_solid_triangle(tri)).

See also each_ghost_triangle and each_solid_triangle.

DelaunayTriangulation.each_solid_triangle — Method

each_solid_triangle(tri::Triangulation)

Given a triangulation tri, returns an iterator over the solid (i.e. non-ghost) triangles.

See also each_ghost_triangle and each_triangle.

DelaunayTriangulation.each_ghost_triangle — Method

each_ghost_triangle(tri::Triangulation)

Given a triangulation tri, returns an iterator over the ghost triangles.

See also each_triangle and each_solid_triangle.

DelaunayTriangulation.contains_triangle — Method

contains_triangle(tri::Triangulation, T) = contains_triangle(T, get_triangles(tri))
contains_triangle(tri::Triangulation, i, j, k) = contains_triangle(i, j, k, get_triangles(tri))

Given a triangulation tri, tests if the triangle T = (i, j, k) is in the triangulation (at least, in triangles). The returned value is a tuple, with the first result being the form of T that is actually in the triangulation (e.g., if T = (i, j, k) but the form of T in tri is (k, i, j), (k, i, j) is returned). The second argument is a Bool, with true if T is contained in tri and false otherwise. If false, the first output is T.

DelaunayTriangulation.construct_positively_oriented_triangle — Method

construct_positively_oriented_triangle(tri::Triangulation, i, j, k)

Given a triangulation tri and indices i, j, k for a proposed triangle, returns a triangle with the indices ordered such that (i, j, k) is not negatively oriented.

DelaunayTriangulation.edge_type — Method

edge_type(tri::Triangulation{P,Ts,I,E}) where {P,Ts,I,E}

Given a triangulation tri, returns the type used for representing individual edges.

DelaunayTriangulation.num_edges — Method

num_edges(tri::Triangulation)

Given a triangulation tri, returns the number of edges. Note that this does not double count edges, e.g. (i, j) and (j, i) do not count as two edges.

DelaunayTriangulation.each_edge — Method

each_edge(tri::Triangulation)

Given a triangulation tri, returns an iterator over the edges in the triangulation. Note that e.g. (i, j) and (j, i) will not be iterated over twice, only one of them would be.

DelaunayTriangulation.get_points — Method

get_points(tri::Triangulation)

Returns the points field from the triangulation tri.

DelaunayTriangulation.get_point — Method

get_point(tri::Triangulation, i...)

Given a triangulation tri and some indices i..., returns the ith point from the triangulation. If i is just a single integer, the result is a Tuple of the coordinates, but if there are multiple integers provided then a Tuple of Tuples of these coordinates is returned, with the jth Tuple the coordinates for the i[j]th point.

It is assumed that whenever i is not an integer, i is meant to be a point, so (getx(i), gety(i)) would be returned in that case. This makes it easier to use some predicates without having to know the index of the point, simply passing the point directly.

If is_boundary_index(i), then instead of returning the ith point of the triangulation, the centroid for the boundary curve corresponding to the boundary index i is returned. See also get_representative_point_coordinates.

DelaunayTriangulation.each_point_index — Method

each_point_index(tri::Triangulation)

Given a triangulation tri, returns an iterator over the indices of the points in tri. This iterator does not include the boundary indices.

DelaunayTriangulation.each_point — Method

each_point(tri::Triangulation)

Given a triangulation tri, returns an iterator over the coordinates of the points in tri. This iterator does not include the centroids of the boundary curves.

DelaunayTriangulation.num_points — Method

num_points(tri::Triangulation)

Given a triangulation tri, returns the number of points.

DelaunayTriangulation.is_boundary_edge — Method

is_boundary_edge(tri::Triangulation, ij)
is_boundary_edge(tri::Triangulation, i, j)

Given a triangulation tri and an edge (i, j), returns true if (i, j) is a boundary edge or false if not.

This predicate takes care for the orientation of the edge: a boundary edge is one that has only two solid vretices, so that get_adjacent(tri, ij) ≤ I(BoundaryIndex).

DelaunayTriangulation.get_boundary_map — Method

get_boundary_map(tri::Triangulation)

Returns the boundary_map field from the triangulation tri.

DelaunayTriangulation.get_boundary_index_ranges — Method

get_boundary_index_ranges(tri::Triangulation)

Returns the boundaryindexranges field from the triangulation tri.

DelaunayTriangulation.is_boundary_triangle — Method

is_boundary_triangle(tri::Triangulation, i, j, k)
is_boundary_triangle(tri::Triangulation, T)

Given a triangulation tri and a triangle T, returns true if T is a boundary triangle or false if not.

Note that this is different from testing if T is a ghost triangle - it is assumed that T is a solid triangle.

DelaunayTriangulation.triangle_orientation — Method

triangle_orientation(tri::Triangulation, i, j, k)
triangle_orientation(tri::Triangulation, T)

Given a triangle T = (i, j, k) and a triangulation tri, computes the orientation of the triangle. In particular, returns:

Certificate.PositivelyOriented: The triangle is positively oriented.

Certificate.Degenerate: The triangle is degenerate, meaning the coordinates are collinear.

Certificate.NegativelyOriented: The triangle is negatively oriented.

A test is also made for the case that is_ghost_triangle(T): If T is a ghost triangle, then the index corresponding to a boundary index points to a centroid, in which case one of the edges has its orientation flipped. This case will also be handled correctly. In case the boundary index corresponds to an interior curve, this flip is not necessary.

DelaunayTriangulation.point_position_relative_to_circumcircle — Method

point_position_relative_to_circumcircle(tri::Triangulation, i, j, k, ℓ)
point_position_relative_to_circumcircle(tri::Triangulation, T, ℓ)

Given a triangle T = (i, j, k), a triangulation tri, and a point with index ℓ, pₗ, computes the position this point pₗ relative to the triangle's circumcircle, making use of point_position_relative_to_circle.

The returned values are:

Certificate.Outside: pₗ is outside of the circumcircle.

Certificate.On: pₗ is on the circumcircle.

Certificate.Inside: pₗ is inside the circumcircle.

A test is also made for the case that is_ghost_triangle(T): When T is a ghost triangle, one of its indices is a boundary index, say i. Since this vertex is treated as being out at infinity, the circumcircle degenerates into the line through the other two vertices. Thus, we test that pₗ is inside this circumcircle by seeing if it is in the oriented outer halfplane defined by the two other vertices. See also point_position_relative_to_oriented_outer_halfplane.

DelaunayTriangulation.point_position_relative_to_line — Method

point_position_relative_to_line(tri::Triangulation, i, j, u)

Given indices i, j, and u corresponding to indices of points in the triangulation tri, corresponding to coordinates say a, b, and p, respectively, computes the position of p relative to the oriented line (a, b).

The returned values are:

Certificate.Left: p is to the left of the line.

Certificate.Collinear: p is on the line.

Certificate.Right: p is to the right of the line.

If is_ghost_edge(i, j), the oriented line (a, b) is flipped since the point corresponding to the boundary index will be a centroid which swaps the orientation.

DelaunayTriangulation.point_position_on_line_segment — Method

point_position_on_line_segment(tri::Triangulation, i, j, u)

Given indices i, j, and u corresponding to indices of points in the triangulation tri, corresponding to coordinates say a, b, and p, respectively, computes the position of p relative to the oriented line segment (a, b), assuming that the three points are collinear.

The returned values are:

Certificate.On: p is on the line segment, meaning between a and b.

Certificate.Degenerate: Either p == a or p == b, i.e. p is one of the endpoints.

Certificate.Left: p is off and to the left of the line segment.

Certificate.Right: p is off and to the right of the line segment.

DelaunayTriangulation.line_segment_intersection_type — Method

line_segment_intersection_type(tri::Triangulation, u, v, i, j)

Given two pairs of indices (u, v) and (i, j), with all indices corresponding to points in the triangulation tri, tests the number of intersections between the two line segments associated with these indices.

Letting p, q, a, and b be the points referred to by u, v, i, and j, respectively, we return:

Certificate.None: The line segments do not meet at any points.

Certificate.Multiple: The closed line segments [p, q] and [a, b] meet in one or several points.

Certificate.Single: The open line segments (p, q) and (a, b) meet in a single point.

Certificate.On: One of the endpoints is on [a, b], but there are no other intersections.

DelaunayTriangulation.point_position_relative_to_triangle — Method

point_position_relative_to_triangle(tri::Triangulation, i, j, k, u)
point_position_relative_to_triangle(tri::Triangulation, T, u)

Given a triangle T = (i, j, k) and another index u, with indices referring to points in the triangulation tri, computes the position of u relative to the triangle.

Letting p, q, a, and b be the points referred to by i, j, k, and u, respectively, we return:

Certificate.Outside: p is outside of the triangle.

Certificate.On: p is on one of the edges.

Certificate.Inside: p is inside the triangle.

If T is a ghost triangle, Certificate.Inside is returned.

DelaunayTriangulation.is_outer_boundary_index — Method

is_outer_boundary_index(tri::Triangulation, i)

Given a triangulation tri and an index i, tests if i is the index of a boundary corresponding to the outermost boundary of the triangulation.

DelaunayTriangulation.is_outer_boundary_node — Method

is_outer_boundary_node(tri::Triangulation, i)

Given a triangulation tri and an index i, tests if i is the index of a node on the outermost boundary of the triangulation.

DelaunayTriangulation.edge_exists — Method

edge_exists(tri::Triangulation, i, j)
edge_exists(tri::Triangulation, ij)

Given a triangulation tri and an edge (i, j), tests if the edge (i, j) is in the triangulation.

DelaunayTriangulation.has_ghost_triangles — Method

has_ghost_triangles(tri::Triangulation)

Given a triangulation tri, tests if the triangulation has ghost triangles.

DelaunayTriangulation.has_boundary_nodes — Method

has_boundary_nodes(tri::Triangulation)

Given a triangulation tri, tests if the triangulation has boundary nodes - these are nodes that are constrained to be there, not those on the convex hull.

DelaunayTriangulation.brute_force_search — Method

brute_force_search(tri::Triangulation, q)

Given a point q, finds the triangle in the triangulation tri containing it by searching over all triangles.

DelaunayTriangulation.jump_and_march — Method

jump_and_march(tri::Triangulation, q;
    m=default_num_samples(num_points(tri)),
    point_indices=each_point_index(tri),
    try_points=(),
    k=select_initial_point(get_points(tri), q; m, point_indices, try_points),
    check_existence::C=Val(has_multiple_segments(tri)),
    rng::AbstractRNG=Random.default_rng())

Using the jump and march algorithm, finds the triangle in the triangulation tri containing the query point q.

Arguments

tri: The Triangulation.

q: The query point.

Keyword Arguments

m=default_num_samples(num_points(pts)): The number of samples to use when sampling an initial point from select_initial_point. Only relevant if k is not specified.

point_indices: The indices for the points. Only relevant if k is not specified.

try_points=(): Extra points to try when sampling an initial point from select_initial_point. Only relevant if k is not specified.

k=select_initial_point(pts, q; m, point_indices, try_points): Where to start the algorithm.

check_existence::C=Val(has_multiple_segments(tri)): Whether to check that the edge exists when using get_adjacent, helping to correct for incorrect boundary indices in the presence of multiple boundary segments. See get_adjacent.

Output

Returns the triangle V containing the query point q.

DelaunayTriangulation.integer_type — Method

integer_type(tri::Triangulation{P,Ts,I}) where {P,Ts,I}

Given a triangulation tri, returns the type used for representing the integers.

DelaunayTriangulation.number_type — Method

number_type(::Triangulation{P}) where {P}

Given a triangulation tri, returns the type used for representing individual coordinates.

DelaunayTriangulation.compute_representative_points! — Method

compute_representative_points!(tri::Triangulation; use_convex_hull=!has_multiple_segments(tri) && num_boundary_edges(get_boundary_nodes(tri)) == 0))

Given a triangulation tri, computes representative points for each region using pole_of_inaccessibility. If you only want to update the point for the convex hull, set use_convex_hull = true - this will only update RepresentativePointList[1].

See also RepresentativePointList.

Warning

While this function computes an appropriate visual center of the polygon represented by the curves, i.e. by joining points, the update functions like update_centroid_after_addition and update_centroid_after_deletion treat the centroid as if it is a mean of all the points. If you need to use the actual visual centers of the polygons, you will need to reuse this function again (as is done at the end of triangulate).

DelaunayTriangulation.clear_empty_features! — Method

clear_empty_features!(tri::Triangulation)

Removes empty keys from the Adjacent and Adjacent2Vertex maps from the triangulation tri, and all points with empty neighbourhoods from the Graph.

See also clear_empty_keys! and clear_empty_points!.

DelaunayTriangulation.find_edge — Method

find_edge(tri::Triangulation, T, ℓ)

Given a triangle T, a triangulation tri, and a point index ℓ which is assumed to correspond to some point on an edge of T, returns the edge (u, v) that the point is on.

DelaunayTriangulation.get_constrained_edges — Method

get_constrained_edges(tri::Triangulation)

Returns the constrained_edges field from the triangulation tri.

DelaunayTriangulation.is_constrained — Method

is_constrained(tri::Triangulation)

Returns true if tri has any constrained edges, and false otherwise.

DelaunayTriangulation.all_boundary_indices — Method

all_boundary_indices(tri::Triangulation)

Returns an iterator over all boundary indices in the triangulation.