Operations

We define some specific operations for acting on Triangulations directly. These are listed below.

DelaunayTriangulation.add_boundary_information! — Method

add_boundary_information!(tri::Triangulation)

Given a Triangulation tri, adds boundary information into tri. In particular, the Adjacent, Adjacent2Vertex, and Graph fields are updated so that e.g. boundary edges map to a corresponding boundary index from Adjacent, boundary indices map to sets of boundary edges from Adjacent2Vertex, and boundary indices map to boundary nodes from Graph.

No values are returned.

DelaunayTriangulation.add_ghost_triangles! — Method

add_ghost_triangles!(tri::Triangulation)

Given a Triangulation tri, adds the ghost triangles into tri. In particular, the triangles, [Adjacent](@ref), and [Adjacent2Vertex`](@ref) fields are updated so that ghost triangles are stored in them.

A ghost triangle is a triangle of the form (i, j, k) where only one of the indices, say i, satisfies is_boundary_index(i).

No values are returned.

DelaunayTriangulation.add_point! — Method

add_point!(tri::Triangulation, new_point;
point_indices=get_vertices(tri),
m=default_num_samples(length(point_indices)),
try_points=(),
initial_search_point=integer_type(tri)(select_initial_point(get_points(tri),
                                                            new_point;
                                                            point_indices,
                                                            m,
                                                            try_points)))

Adds the point new_point into the triangulation.

This function will not update the convex hull - if you need it to be corrected, you could use e.g. convex_hull!. You should also be careful with using this if you have deleted ghost triangles.

Arguments

tri::Triangulation: The triangulation.

new_point: The index of the point in get_points(tri) to add.

Keyword Arguments

point_indices=get_vertices(tri): The currently inserted points in tri.

m=default_num_samples(length(point_indices)): How many points to sample.

try_points(): Points to consider when sampling for point location.

initial_search_point=integer_type(tri)(select_initial_point(get_points(tri), new_point; point_indices, m, try_points))): Where to start the point location.

Outputs

There are no outputs, but tri is updated in-place.

DelaunayTriangulation.add_triangle! — Method

add_triangle!(tri::Triangulation, T; update_ghost_edges=false)
add_triangle!(tri::Ts, u::Integer, v::Integer, w::Integer; update_ghost_edges=false) where {Ts<:Triangulation}

Given a triangle T, or a triple of integers u, v, w, adds the triangle into the Triangulation type tri, updating the Adjacent, Adjacent2Vertex, Graph, and triangles fields.

To update the ghost triangles directly, set update_ghost_edges=true. The other parts of the boundary information will be handled, though.

No values are returned.

DelaunayTriangulation.delete_ghost_triangles! — Method

delete_ghost_triangles!(tri::Triangulation)

Given a Triangulation tri, deletes the ghost triangles from tri. In particular, the triangles, Adjacent, and Adjacent2Vertex fields are updated so that ghost triangles are no longer stored in them.

A ghost triangle is a triangle of the form (i, j, k) where one of the indices, say i, satisfies is_boundary_index(i).

No values are returned.

DelaunayTriangulation.delete_triangle! — Method

delete_triangle!(tri::Triangulation, T)
delete_triangle!(tri::Ts, u::Integer, v::Integer, w::Integer) where {Ts<:Triangulation}

Given a triangle T, or a triple of integers u, v, w, deletes the triangle from the Triangulation type tri, updating the Adjacent, Adjacent2Vertex, Graph, and triangles fields.

No values are returned.