Point Location

The most expensive step for building a Delaunay triangulation is the point location step, wherein we need to find a triangle that contains a given point. The code for this turns out to be very complicated so that we can correctly handle points outside of the domain, inside interior holes, collinear with other points, on the corner, etc. The main function that handles all of this is jump_and_march, derived from the jump-and-march algorithm of Mücke, Saias, and Zhu (1999). We also provide a method that simply searches over all triangles, given below.

DelaunayTriangulation.brute_force_search — Function

brute_force_search(tri::Triangulation, q)

Given a point q, finds the triangle in the triangulation tri containing it by searching over all triangles.

brute_force_search(T, r, pts, boundary_map)

Given a collection of triangles T, a point r, a collection of points pts, and a boundary map handling the mapping of boundary indices from construct_boundary_map, returns the triangle in T containing r by searching over all triangles.

The Main Method

Below we list docstrings for the main jump and march algorithm.

DelaunayTriangulation.jump_and_march — Function

jump_and_march(tri::Triangulation, q;
    m=default_num_samples(num_points(tri)),
    point_indices=each_point_index(tri),
    try_points=(),
    k=select_initial_point(get_points(tri), q; m, point_indices, try_points),
    check_existence::C=Val(has_multiple_segments(tri)),
    rng::AbstractRNG=Random.default_rng())

Using the jump and march algorithm, finds the triangle in the triangulation tri containing the query point q.

Arguments

tri: The Triangulation.

q: The query point.

Keyword Arguments

m=default_num_samples(num_points(pts)): The number of samples to use when sampling an initial point from select_initial_point. Only relevant if k is not specified.

point_indices: The indices for the points. Only relevant if k is not specified.

try_points=(): Extra points to try when sampling an initial point from select_initial_point. Only relevant if k is not specified.

k=select_initial_point(pts, q; m, point_indices, try_points): Where to start the algorithm.

check_existence::C=Val(has_multiple_segments(tri)): Whether to check that the edge exists when using get_adjacent, helping to correct for incorrect boundary indices in the presence of multiple boundary segments. See get_adjacent.

Output

Returns the triangle V containing the query point q.

jump_and_march(pts, adj, adj2v, graph::Graph{I}, boundary_index_ranges, boundary_map, q;
    m=default_num_samples(num_points(pts)),
    point_indices=each_point_index(pts),
    try_points=(),
    k=select_initial_point(pts, q; m, point_indices, try_points),
    TriangleType::Type{V}=NTuple{3,I},
    check_existence::C=Val(has_multiple_segments(boundary_map)),
    rng::AbstractRNG = Random.default_rng()) where {I,V,C}

Using the jump and march algorithm, finds the triangle in the triangulation containing the query point q.

If your triangulation does not have ghost triangles, and the point q is outside of the triangulation, this function may fail to terminate. You may like to add ghost triangles in this case (using add_ghost_triangles!), noting that there is no actual triangle that q is inside when it is outside of the triangulation unless ghost triangles are present.

Arguments

pts: The collection of points.

adj: The Adjacent map.

graph::Graph{I}: The Graph.

boundary_index_ranges: The Dict mapping boundary indices to ranges from construct_boundary_index_ranges.

boundary_map: The boundary map from construct_boundary_map handling the mapping of boundary indices.

q: The query point.

Keyword Arguments

m=default_num_samples(num_points(pts)): The number of samples to use when sampling an initial point from select_initial_point. Only relevant if k is not specified.

point_indices: The indices for the points. Only relevant if k is not specified.

try_points=(): Extra points to try when sampling an initial point from select_initial_point. Only relevant if k is not specified.

k=select_initial_point(pts, q; m, point_indices, try_points): Where to start the algorithm.

TriangleType::Type{V}=NTuple{3,I}: The type used for representing the triangles.

check_existence::C=Val(has_multiple_segments(boundary_map)): Whether to check that the edge exists when using get_adjacent, helping to correct for incorrect boundary indices in the presence of multiple boundary segments. See get_adjacent.

Output

Returns the triangle V containing the query point q.

The variant of the jump and march algorithm for points outside of the triangle is also accessed via the jump_and_march, calling into exterior_jump_and_march:

DelaunayTriangulation.exterior_jump_and_march — Function

exterior_jump_and_march(pts, adj::Adjacent{I,E}, boundary_index_ranges, boundary_map, k, q, check_existence::V=Val(has_multiple_segments(boundary_map))) where {I,E}

Given a collection of points pts, an Adjacent map adj, a Dict mapping boundary indices to ranges from construct_boundary_index_ranges, a boundary map from construct_boundary_map, a vertex k, and a point q outside of the triangulation, returns the edge (i, j) such that the ghost triangle (i, j, -1) contains q.

The check_existence argument is used to check that the edge exists when using get_adjacent, helping to correct for incorrect boundary indices in the presence of multiple boundary segments. See get_adjacent.

The result is meaningless if q is inside of the triangulation.

You should not need to call into this method directly.

The core complexity of the algorithm comes from having to find the direction of the point from an initial search point. The docstrings for some of these initialisers are given below.

DelaunayTriangulation.select_initial_point — Function

select_initial_point(pts, q;
point_indices=each_point_index(pts),
m=default_num_samples(length(point_indices)),
try_points=(),
rng::AbstractRNG = Random.default_rng())

Given a collection of points and a point q, select m random points from pts and return the index i that corresponds a point get_point(pts, i) closest to q, out of all indices in point_indices and try_points. The indices are randomly selected from point_indices, but you can guarantee points to try using try_points.

If q is an integer, then the point compared to becomes get_point(pts, q).

DelaunayTriangulation.select_initial_triangle_interior_node — Function

select_initial_triangle_interior_node(pts, adj::Adjacent, adj2v::Adjacent2Vertex, boundary_map, k, q, boundary_index_ranges, check_existence::V = Val(has_multiple_segments(boundary_map)), rng::AbstractRNG = Random.default_rng()) where {V}

Selects an initial triangle for the jump-and-march algorithm, starting from a point with index k and searching for the point q. The point q should be contained within the outermost boundary (interior holes are fine), and k should not be a point on the outer boundary.

Arguments

pts: The collection of points.

adj: The Adjacent map.

adj2v: The Adjacent2Vertex map.

boundary_map: The map taking boundary indices to their boundary segment location. See construct_boundary_map.

k: The index of the point in pts that we are starting at.

q: The point being searched for.

boundary_index_ranges: The Dict handling the mapping of boundary indices from construct_boundary_index_ranges.

check_existence::V=Val(has_multiple_segments(boundary_map))): Checks for different possible boundary indices when there are multiple segments. See get_adjacent.

Outputs

p: The kth point in pts.

i, j: These are indices defining the edge of a triangle including the point p, such that i is to the left of the line pq and j is to the right of pq.

pᵢ, pⱼ: The points in pts corresponding to the indices in i and j, respectively.

DelaunayTriangulation.check_for_intersections_with_adjacent_boundary_edges — Function

check_for_intersections_with_adjacent_boundary_edges(pts, adj::Adjacent{I,E}, boundary_index_ranges, boundary_map, k, q, check_existence::V = Val(has_multiple_segments(boundary_map))) where {I,E,V}

Given a collection of points pts, an Adjacent map adjacent, a list of boundary index ranges from construct_boundary_index_ranges, a boundary map from construct_boundary_map, an outer boundary point k, and a query point q, checks the edges adjacent to k along the outer boundary for q.

The check_existence argument is the same keyword argument from get_adjacent, and is needed when you use multiple segments in your boundary.

See also search_down_adjacent_boundary_edges, which uses this function to determine an initial direction to search along a straight boundary in case q is collinear with it.

The possible returned values are:

(Certificate.Outside, Certificate.Outside, k)

Here, the point q is not collinear with either of the adjacent boundary edges.

(Certificate.Right, C, r), where C is either Certificate.On or Certificate.Right and r is the vertex to the right of k, i.e. get_adjacent(adj, k, -1)

This output means that q is collinear with the edge to the right of k. C = Certificate.On means it is on this edge, while C = Certificate.Right means that q is to the right of this edge.

(Certificate.Left, C, ℓ), where C is either Certificate.On or Certificate.Left and ℓ is the vertex to the left of k, i.e. get_adjacent(adj, -1, k)

This output means that q is collinear with the edge to the left of k. C = Certificate.On means it is on this edge, while C = Certificate.Left means that q is to the left of this edge.

In these two outputs above, C could also mean Certificate.Degenerate, which means that q is get_point(pts, r) or get_point(pts, ℓ), respectively.

In addition to these three returned values, the fourth and fifth returned values are (right_cert, left_cert), which give the position of q relative to the edges (p, p_right) and (p, p_left), respectively, where p_right is the point on the boundary to the right and p_left is the point on the boundary to the left of p. These returned values are useful in case we need to go to check_for_intersections_with_interior_edges_adjacent_to_boundary_node, since we can reuse these certificates.

DelaunayTriangulation.search_down_adjacent_boundary_edges — Function

search_down_adjacent_boundary_edges(pts, adj::Adjacent{I,E}, boundary_index_ranges, boundary_map, k, q, direction, q_pos, next_vertex, check_existence::V=Val(has_multiple_segments(boundary_map))) where {I,E,V}

Given a collection of points pts, an Adjacent map adj, a list of boundary index ranges from construct_boundary_index_ranges, a boundary map from construct_boundary_map, an outer boundary index k for a point in pts, a point q being searched for, a direction giving the direction of q from get_point(pts, k), a certificate for the position of q from this point, and the next vertex in the direction of q (these last three arguments coming from check_for_intersections_with_adjacent_boundary_edges), walks down the edges that q is collinear with until an edge is found that q is on or until finding that q is outside of the triangulation.

The returned value takes the form (cert, u, v, w), with cert = Certificate.On if q is on the edge (u, v) and cert = Certificate.Outside if q is outside of the triangulation. If is_on(cert), then (u, v, w) is a positively oriented triangle with q on the edge (u, v). Otherwise, (u, v, w) is a ghost triangle that should be close to q.

The check_existence argument is the same keyword argument from get_adjacent, and is needed when you use multiple segments in your boundary.

Note that this function relies on the assumption that the geometry is convex.

DelaunayTriangulation.check_for_intersections_with_interior_edges_adjacent_to_boundary_node — Function

check_for_intersections_with_interior_edges_adjacent_to_boundary_node(pts, adj::Adjacent{I,E}, graph::Graph{I}, boundary_index_ranges, boundary_map, k, q, right_cert, left_cert, check_existence::V=Val(has_multiple_segments(boundary_map))) where {I,E,V}

Checks if the line connecting the kth point of pts to q intersects any of the edges neighbouring the boundary node k.

This function should only be used after check_for_intersections_with_adjacent_boundary_edges, and currently is only guaranteed to work on convex geometries.

Arguments

pts: The collection of points.

adj::Adjacent{I,E}: The Adjacent map.

graph::Graph{I}: The Graph.

boundary_index_ranges: The boundary index range mapping from construct_boundary_index_ranges.

boundary_map: The map that handles the mapping of boundary indices to boundary segments. Sse construct_boundary_map.

k: The boundary node.

q: The point we are searching for.

right_cert: A certificate giving the position of q to the right of the kth point. This comes from check_for_intersections_with_adjacent_boundary_edges.

left_cert: A certificate giving the position of q to the left of the kth point. This comes from check_for_intersections_with_adjacent_boundary_edges.

check_existence::V=Val(has_multiple_segments(boundary_nodes))): Checks for different possible boundary indices when there are multiple segments. See get_adjacent.

Outputs

There are several possible forms for the returned values. These are listed below, letting p be the kth point, pᵢ the point corresponding to the index i, and pⱼ the point corresponding to the index j:

(i, j, Certificate.Single, Certificate.Outside)

The line pq intersects the edge pᵢpⱼ, and (j, i, k) is a positively oriented triangle. In particular, pᵢ is left of pq and pⱼ is right of pq.

(i, j, Certificate.None, Certificate.Inside)

The point q is inside the positively oriented triangle (i, j, k).

(zero(I), zero(I), Cert.None, Cert.Outside)

The point q is outside of the triangulation.

(i, j, Cert.On, Cert.Inside)

The point q is on the edge pᵢpⱼ, and so is inside the positively oriented triangle (i, j, k).

(i, j, Cert.Right, Cert.Outside)

The point q is collinear with the edge pᵢpⱼ, but is off of it and further into the triangulation.

Basic Description of the Algorithm

Let us give a basic description of what jump_and_march does.

First, using select_initial_point, an initial point to start the algorithm is selected. This function samples some number $m$ of points, and then selects the point that is closest to the query point $q$.

Let the initially selected point be $p_k$. We break the discussion into two cases, where $p_k$ is an interior point and $p_k$ is a point on the boundary.

If $p_k$ is not a point on the boundary, then it is possible to completely rotate around the point searching for a triangle such that line $\overrightarrow{p_kq}$ intersects an edge of the triangle. This will give us an edge $e_{ij}$ that $\overrightarrow{p_kq}$ intersects, and we will put $p_i$ to the left of $\overrightarrow{p_kq}$ and $p_j$ to the right. The function that handles this selection is select_initial_triangle_interior_node, which starts by handling the case of collinear edges and then rotates around.

Now suppose that $p_k$ is a point on the outer boundary. There are several possibilities in this case. First, $\overrightarrow{p_kq}$ might intersect a neighbouring boundary edge; secondly, $\overrightarrow{p_kq}$ could intersect a neighbouring interior edge; thirdly, $\overrightarrow{p_kq}$ could point away from the boundary, meaning $q$ is outside of the boundary. The function starts by checking the neighbouring boundary edges, done via check_for_intersections_with_adjacent_boundary_edges, making use of get_right_boundary_node and get_left_boundary_node to obtain the neighbouring boundary nodes. If we find that $\overrightarrow{p_kq}$ does intersect a neighbouring boundary edge, then we can search down adjacent boundary edges via search_down_adjacent_boundary_edges until we either find an edge that $q$ is on, or until we identify that $q$ is outside of the triangulation – this function assumes the domain is convex, as triangulations being built are. If $q$ is outside of the triangulation, then we can use the exterior variant of the jump and march algorithm, exterior_jump_and_march, to find the ghost triangle containing $q$. This functon simply rotates around the boundary until we find two ghost edges enclosing the point $q$. Now let us assume we did not find a neighbouring boundary edge that intersects $\overrightarrow{p_kq}$. When this happens, we need to check the neighbouring interior edges, done via check_for_intersections_with_interior_edges_adjacent_to_boundary_node, which just rotates around the edges until we find an intersection. If we have still not found any intersection then, again assuming convexity, $q$ must be outside of the triangulation and so we use exterior_jump_and_march.

Now, if the algorithm is still going, then we need to start marching along the triangulation from $p_k$ towards $q$. The idea is to keep marching, keeping $p_i$ and $p_j$ to the left and right of $\overrightarrow{p_kq}$, respectively, until we find a case where $p_ip_jq$ is no longer a positively oriented triangle. When this happens, it must mean that we have passed $q$, and so we have found the triangle.

Let us describe how this marching is done in more detail. First, we need to be careful of boundary indices, first checking for an outer boundary index. If we have found an outer boundary index, then we have marched into the boundary, and so $q$ will be outside of the domain, meaning we go into exterior_jump_and_march. If this is not the case, then we march using get_adjacent to step onto the next triangle from a given edge $e_{ij}$. Then, checking the positions of the new points relative to $\overrightarrow{p_kq}$ and rearranging accordingly, we can step forward. We keep doing this until we get a negatively oriented triangle $p_ip_jq$. Unfortunately, this loop could terminate even if $q$ is not in the found triangle, which can occasionally happen if $p_ip_jq$ is a degenerate triangle. In this case, we just restart the jump and march algorithm. This latter worry is a very rare concern and does not alter the runtime in any significant manner.