Predicates

The predicates that we use in this package are all built from ExactPredicates.jl, avoiding degeneracies from predicates owing to floating point arithmetic. The results from predicates are based on certificates, coming from a Certificate type defined with EnumX.jl. The definition of this is below.

DelaunayTriangulation.Certificate — Module

Certificate

An Enum type that represents results from a geometric predicate. Below we provide a list of available certificates, along with the function that can be used for testing if a given Certificate matches that certificate.

Inside: is_inside

Degenerate: is_degenerate

Outside: is_outside

On: is_on

Left: is_left

Right: is_right

PositivelyOriented: is_positively_oriented

NegativelyOriented: is_negatively_oriented

Collinear: is_collinear

None: is_none or has_no_intersections

Single: is_single or has_one_intersection

Multiple: is_multiple or has_multiple_intersections

Touching: is_touching

General

Below we list some general predicates.

DelaunayTriangulation.orient_predicate — Function

orient_predicate(p, q, r)

Returns ExactPredicates.orient(p, q, r).

DelaunayTriangulation.incircle_predicate — Function

incircle_predicate(a, b, c, p)

Returns ExactPredicates.incircle(a, b, c, p).

DelaunayTriangulation.sameside_predicate — Function

sameside_predicate(a, b, p)

Returns ExactPredicates.sameside(p, a, b).

(The difference in the argument order is to match the convention that the main point being tested is the last argument.)

DelaunayTriangulation.meet_predicate — Function

meet_predicate(p, q, a, b)

Returns ExactPredicates.meet(p, q, a, b).

DelaunayTriangulation.triangle_orientation — Method

triangle_orientation(p, q, r)

Given a triangle with coordinates (p, q, r), computes its orientation, returning:

Certificate.PositivelyOriented: The triangle is positively oriented.

Certificate.Degenerate: The triangle is degenerate, meaning the coordinates are collinear.

Certificate.NegativelyOriented: The triangle is negatively oriented.

See also orient_predicate.

DelaunayTriangulation.point_position_relative_to_circle — Function

point_position_relative_to_circle(a, b, c, p)

Given a circle through the coordinates (a, b, c), assumed to be positively oriented, computes the position of p relative to the circle. In particular, returns:

Certificate.Inside: p is inside the circle.

Certificate.On: p is on the circle.

Certificate.Outside: p is outside the triangle.

See also incircle_predicate.

DelaunayTriangulation.point_position_relative_to_line — Method

point_position_relative_to_line(a, b, p)

Given a point p and the oriented line (a, b), computes the position of p relative to the line, returning:

Certificate.Left: p is to the left of the line.

Certificate.Collinear: p is on the line.

Certificate.Right: p is to the right of the line.

See also orient_predicate.

DelaunayTriangulation.point_position_on_line_segment — Method

point_position_on_line_segment(a, b, p)

Given a point p and the line segment (a, b), assuming p to be collinear with a and b, computes the position of p relative to the line segment. In particular, returns:

Certificate.On: p is on the line segment, meaning between a and b.

Certificate.Degenerate: Either p == a or p == b, i.e. p is one of the endpoints.

Certificate.Left: p is off and to the left of the line segment.

Certificate.Right: p is off and to the right of the line segment.

See also sameside_predicate.

DelaunayTriangulation.line_segment_intersection_type — Method

line_segment_intersection_type(p, q, a, b)

Given the coordinates (p, q) and (a, b) defining two line segments, tests the number of intersections between the two segments. In particular, we return:

Certificate.None: The line segments do not meet at any points.

Certificate.Multiple: The closed line segments [p, q] and [a, b] meet in one or several points.

Certificate.Single: The open line segments (p, q) and (a, b) meet in a single point.

Certificate.Touching: One of the endpoints is on [a, b], but there are no other intersections.

See also meet_predicate.

DelaunayTriangulation.point_position_relative_to_triangle — Method

point_position_relative_to_triangle(a, b, c, p)

Given a positively oriented triangle with coordinates (a, b, c), computes the position of p relative to the triangle. In particular, returns:

Certificate.Outside: p is outside of the triangle.

Certificate.On: p is on one of the edges.

Certificate.Inside: p is inside the triangle.

DelaunayTriangulation.point_position_relative_to_oriented_outer_halfplane — Function

point_position_relative_to_oriented_outer_halfplane(a, b, p)

Given an edge with coordinates (a, b) and a point p, tests the position of p relative to the oriented outer halfplane defined by (a, b). The oriented outer halfplane is the union of the open halfplane defined by the region to the left of the oriented line (a, b), and the open line segment (a, b). The returned values are:

Cert.Outside: p is outside of the oriented outer halfplane, meaning to the right of the line (a, b) or collinear with a and b but not on the line segment (a, b).

Cert.On: p is on the open line segment (a, b).

Cert.Inside: p is inside of the oriented outer halfplane, meaning to the left of the line (a, b).

Boundaries and Ghosts

Below we list some predicates for working with boundaries and ghost triangles.

DelaunayTriangulation.is_boundary_index — Function

is_boundary_index(i::I) where {I}

Given an index i, returns i ≤ I(BoundaryIndex).

DelaunayTriangulation.is_boundary_edge — Method

is_boundary_edge(ij, adj::Adjacent)
is_boundary_edge(i, j, adj::Adjacent{I,E}) where {I,E}

Given an edge (i, j) and an adjacent map adj, returns true if (i, j) is a boundary edge and false otherwise.

Note that the orientation of (i, j) is important: even if (i, j) is an edge on the boundary, if there is a triangle (i, j, k) in the triangulation then (i, j) is not a boundary edge but (j, i) would be.

DelaunayTriangulation.is_boundary_triangle — Method

is_boundary_triangle(i, j, k, adj)
is_boundary_triangle(T, adj)

Given a triangle T = (i, j, k) and an adjacent map adj, returns true if T is a boundary triangle. A boundary triangle is still part of the triangulation, but it has at least one edge that forms part of the boundary (so that at least one of is_boundary_edge(j, i), is_boundary_edge(k, j), and is_boundary_edge(i, k) is true).

DelaunayTriangulation.is_ghost_edge — Function

is_ghost_edge(i, j)

Given an edge (i, j), returns true if (i, j) is a ghost edge. A ghost edge is an edge in which either is_boundary_index(i) or is_boundary_index(j) is true.

DelaunayTriangulation.is_ghost_triangle — Function

is_ghost_triangle(i, j, k)
is_ghost_triangle(T)

Given a triangle T = (i, j, k), returns true if T is a ghost triangle and false otherwise. A ghost triangle is one in which any of the vertices (i, j, k) are a boundary index, as tested via is_boundary_index.

DelaunayTriangulation.is_interior_curve — Function

is_interior_curve(i)
is_interior_curve(i, boundary_map)

Given an index i, tests if the curve is an interior curve, i.e. if i > 1. If a map boundary_map is provided, i should be a boundary map so that is_interior_curve(j) is tested, where j = get_curve_index(boundary_map, i).

DelaunayTriangulation.is_outer_boundary_index — Method

is_outer_boundary_index(i, boundary_map)

Given an index i, tests if the index is a boundary index referring to the outermost boundary, making use of the boundary_map from construct_boundary_map.

DelaunayTriangulation.is_outer_ghost_triangle — Function

is_outer_ghost_triangle(i, j, k, boundary_map)

Given a ghost triangle (i, j, k) and a boundary map boundary_map taking boundary indices to their location in the boundary node array, tests if (i, j, k) is a ghost triangle on the outermost boundary (true) or on an interior boundary (false)`.

DelaunayTriangulation.is_outer_ghost_edge — Function

is_outer_ghost_edge(i, j, boundary_map)

Given a ghost edge (i, j) and a boundary map boundary_map taking boundary indices to their location in the boundary node array, tests if (i, j) is a ghost edge on the outermost boundary (true) or on an interior boundary (false)`.

DelaunayTriangulation.is_outer_boundary_node — Method

is_outer_boundary_node(i, graph::Graph{I}, boundary_index_ranges) where {I}

Given a node index i, a graph::Graph, a Dict from construct_boundary_index_ranges, returns true if i corresponds to a node on the outermost boundary, and false otherwise.

DelaunayTriangulation.edge_exists — Method

edge_exists(i::I) where {I}

Returns i ≠ I(DefaultAdjacentValue).

DelaunayTriangulation.edge_exists — Method

edge_exists(ij, adj::Adjacent{I,E}) where {I,E}

Given an edge ij and an Adjacent map adj, tests if the edge exists in the corresponding triangulation.

DelaunayTriangulation.has_ghost_triangles — Method

has_ghost_triangles(adj::Adjacent{I,E}, adj2v) where {I,E}

Given an Adjacent map adj and an Adjacent2Vertex map adj2v, tests if the corresponding triangulation contains ghost triangles.

Index and Ghost Handling

Below we list methods for working with predicates that are used when we provide indices for points rather than points directly.

DelaunayTriangulation.triangle_orientation — Method

triangle_orientation(i, j, k, pts, boundary_map)
triangle_orientation(T, pts, boundary_map)

Given a triangle T = (i, j, k), with indices corresponding to points in pts, computes the orientation of the triangle. In particular, returns:

Certificate.PositivelyOriented: The triangle is positively oriented.

Certificate.Degenerate: The triangle is degenerate, meaning the coordinates are collinear.

Certificate.NegativelyOriented: The triangle is negatively oriented.

A test is also made for the case that is_ghost_triangle(T): If T is a ghost triangle, then the index corresponding to a boundary index points to a centroid, in which case one of the edges has its orientation flipped. This case will also be handled correctly. In case the boundary index corresponds to an interior curve, this flip is not necessary.

The mapping of a boundary index is handled via the boundary_map argument - see construct_boundary_map.

DelaunayTriangulation.point_position_relative_to_circumcircle — Method

point_position_relative_to_circumcircle(i, j, k, ℓ, pts, boundary_map)
point_position_relative_to_circumcircle(T, ℓ, pts, boundary_map)

Given a triangle T = (i, j, k), with indices corresponding to points in pts, and a point with index ℓ, pₗ, computes the position this point pₗ relative to the triangle's circumcircle, making use of point_position_relative_to_circle.

The returned values are:

Certificate.Outside: pₗ is outside of the circumcircle.

Certificate.On: pₗ is on the circumcircle.

Certificate.Inside: pₗ is inside the circumcircle.

A test is also made for the case that is_ghost_triangle(T): When T is a ghost triangle, one of its indices is a boundary index, say i. Since this vertex is treated as being out at infinity, the circumcircle degenerates into the line through the other two vertices. Thus, we test that pₗ is inside this circumcircle by seeing if it is in the oriented outer halfplane defined by the two other vertices. See also point_position_relative_to_oriented_outer_halfplane.

The mapping of a boundary index is handled via the boundary_map argument - see construct_boundary_map.

DelaunayTriangulation.point_position_relative_to_line — Method

point_position_relative_to_line(i, j, u, pts, boundary_map)

Given indices i, j, and u corresponding to indices of points in the collection of points pts, corresponding to coordinates say a, b, and p, respectively, computes the position of p relative to the oriented line (a, b).

The returned values are:

Certificate.Left: p is to the left of the line.

Certificate.Collinear: p is on the line.

Certificate.Right: p is to the right of the line.

If is_outer_ghost_edge(i, j, boundary_map), the oriented line (a, b) is flipped since the point corresponding to the boundary index will be a centroid which swaps the orientation.

The mapping of a boundary index is handled via the boundary_map argument - see construct_boundary_map.

DelaunayTriangulation.point_position_on_line_segment — Method

point_position_on_line_segment(i, j, u, pts)

Given indices i, j, and u corresponding to indices of points in the collection of points pts, corresponding to coordinates say a, b, and p, respectively, computes the position of p relative to the oriented line segment (a, b), assuming that the three points are collinear.

The returned values are:

Certificate.On: p is on the line segment, meaning between a and b.

Certificate.Degenerate: Either p == a or p == b, i.e. p is one of the endpoints.

Certificate.Left: p is off and to the left of the line segment.

Certificate.Right: p is off and to the right of the line segment.

DelaunayTriangulation.line_segment_intersection_type — Method

line_segment_intersection_type(u, v, i, j, pts)

Given two pairs of indices (u, v) and (i, j), with all indices corresponding to points in pts, tests the number of intersections between the two line segments associated with these indices.

Letting p, q, a, and b be the points referred to by u, v, i, and j, respectively, we return:

Certificate.None: The line segments do not meet at any points.

Certificate.Multiple: The closed line segments [p, q] and [a, b] meet in one or several points.

Certificate.Single: The open line segments (p, q) and (a, b) meet in a single point.

Certificate.On: One of the endpoints is on [a, b], but there are no other intersections.

DelaunayTriangulation.point_position_relative_to_triangle — Method

point_position_relative_to_triangle(i, j, k, u, pts, boundary_map)
point_position_relative_to_triangle(T, u, pts, boundary_map)

Given a triangle T = (i, j, k) and another index u, with indices referring to points in pts, computes the position of u relative to the triangle.

Letting p, q, a, and b be the points referred to by i, j, k, and u, respectively, we return:

Certificate.Outside: p is outside of the triangle.

Certificate.On: p is on one of the edges.

Certificate.Inside: p is inside the triangle.

If necessary, the mapping of boundary indices is handled via the boundary map argument from construct_boundary_map.

DelaunayTriangulation.point_position_relative_to_triangle — Method

point_position_relative_to_triangle(a, b, c, p)

Given a positively oriented triangle with coordinates (a, b, c), computes the position of p relative to the triangle. In particular, returns:

Certificate.Outside: p is outside of the triangle.

Certificate.On: p is on one of the edges.

Certificate.Inside: p is inside the triangle.