Plotting

DelaunayTriangulation.triplot! — Function

triplot(!)(points, triangles, boundary_nodes, convex_hull, args...; kwargs...)
triplot(!)(tri::Triangulation, args...; kwargs...)

Plots a triangulation. The available attributes are:

markersize=11

Markersize for the points.

show_ghost_edges=false

Whether to plot interpretations of the ghost edges. If the ghost edges are on the outermost boundary, then a ghost edge is a line starting at the solid vertex and parallel with the vertex of the domain, pointing outwards. If the ghost edge is an interior ghost edge, then an edge is shown that connects with the centroid of the corresponding interior.

recompute_centers=true

Only relevant if show_ghost_edges. This will recompute the representative centers for each region if needed.

show_all_points=false

Whether to plot all points. If false, only plots those that correspond to a solid triangle.

point_color=:red

Only relevant if show_all_points. Colour to use for the points.

strokecolor=:black

Colour to use for the edges of the triangles.

triangle_color=(:white, 0.0)

Colour to use for the triangles.

ghost_edge_color=:blue

Only relevant if show_ghost_edges. Colour to use for the ghost edges.

ghost_edge_linewidth=1

Only relevant if show_ghost_edges. Linewidth to use for the ghost edges.

strokewidth=1

The width of the triangle edges.

ghost_edge_extension_factor=10.0

Only relevant if show_ghost_edges. For the outermost ghost edges, we extrapolate a line to decide its extent away from the boundary. This is the factor used.

plot_convex_hull=true

Whether to also plot the convex hull of the points.

`convexhullcolor=:grey'

Only relevant if plot_convex_hull. The linecolor to use for the convex hull.

convex_hull_linestyle=:dash

Only relevant if plot_convex_hull. The linestyle to use for the convex hull.

convex_hull_linewidth=2

Only relevant if plot_convex_hull. The linewidth to use for the convex hull.