Other Utilities

Here are some docstrings for other utility functions.

DelaunayTriangulation.get_right_boundary_node — Method

get_right_boundary_node(adj::Adjacent{I,E}, k, boundary_index, boundary_index_ranges, check_existence::C) where {I,E,C}

Given an Adjacent map, a boundary node index k, a boundary_index corresponding to the curve, and boundary_index_ranges from construct_boundary_index_ranges, returns the node on the boundary to the right of k. check_existence can be used if you need to check over all boundary indices, in case there are multiple segments and thus multiple possible boundary indices on the boundary.

See also get_left_boundary_node.

DelaunayTriangulation.get_left_boundary_node — Method

get_left_boundary_node(adj::Adjacent{I,E}, k, boundary_index, boundary_index_ranges, check_existence::C) where {I,E,C}

Given an Adjacent map, a boundary node index k, a boundary_index corresponding to the curve, and boundary_index_ranges from construct_boundary_index_ranges, returns the node on the boundary to the left of k. check_existence can be used if you need to check over all boundary indices, in case there are multiple segments and thus multiple possible boundary indices on the boundary.

See also get_right_boundary_node.

DelaunayTriangulation.find_edge — Method

find_edge(T, points, ℓ)

Given a triangle T and a set of points points, with the ℓth point of points on an edge of T, returns the edge (u, v) that the point is on.

DelaunayTriangulation.choose_uvw — Method

choose_uvw(e1, e2, e3, u, v, w)

Choose values for (u, v, w) based on the Booleans (e1, e2, e3), assuming only one is true. The three cases are:

If e1, returns (u, v, w).

If e2, returns (v, w, u).

If e3, returns (w, u, v).

DelaunayTriangulation.is_circular — Function

is_circular(A)

Tests if A[begin] == A[end].

DelaunayTriangulation.circular_equality — Function

circular_equality(A, B)

Tests if the arrays A and B are equal up to a circular shift.

DelaunayTriangulation.get_surrounding_polygon — Method

get_surrounding_polygon(adj::Adjacent{I,E}, graph::Graph, u, boundary_index_ranges, check_existence::C; skip_boundary_indices=false) where {I,E,C}

Given an Adjacent map, a Graph, a vertex u, boundary_index_ranges from construct_boundary_index_ranges for handling the case where u is on the boundary, returns a vector S which gives a counter-clockwise sequence of the neighbours of u. check_existence can be used if you need to check over all boundary indices when u is a boundary node, in case there are multiple segments and thus multiple possible boundary indices on the boundary.

When u is an outer boundary index, the returned polygon is clockwise.

When u is a boundary vertex and you do not have ghost triangles, then this function may return an invalid polygon.

If you want to remove all boundary indices from the result at the end, set skip_boundary_indices=true.